@@ -1,6 +1,6 @@
 module github.com/reeflective/readline
 
-go 1.19
+go 1.20
 
 require (
 	github.com/google/go-cmp v0.5.8
@@ -8,3 +8,5 @@ require (
 	golang.org/x/sys v0.3.0
 	golang.org/x/term v0.3.0
 )
+
+require github.com/rivo/uniseg v0.4.4
