@@ -10,21 +10,18 @@ import (
 
 func (rl *Shell) completionCommands() commands {
 	return map[string]func(){
-		"complete":               rl.completeWord,        // complete-word
-		"possible-completions":   rl.possibleCompletions, // list-choices
+		"complete":               rl.completeWord,
+		"possible-completions":   rl.possibleCompletions,
 		"insert-completions":     rl.insertCompletions,
 		"menu-complete":          rl.menuComplete,
-		"menu-complete-backward": rl.menuCompleteBackward, // reverse-menu-complete
+		"menu-complete-backward": rl.menuCompleteBackward,
 		"delete-char-or-list":    rl.deleteCharOrList,
 
-		"expand-or-complete":        rl.expandOrComplete,
-		"menu-expand-or-complete":   rl.menuExpandOrComplete,
-		"menu-complete-next-tag":    rl.menuCompleteNextTag,
-		"menu-complete-prev-tag":    rl.menuCompletePrevTag,
-		"accept-and-menu-complete":  rl.acceptAndMenuComplete,
-		"accept-completion-or-line": rl.acceptCompletionOrLine,
-		"vi-registers-complete":     rl.viRegistersComplete,
-		"menu-incremental-search":   rl.menuIncrementalSearch,
+		"menu-complete-next-tag":   rl.menuCompleteNextTag,
+		"menu-complete-prev-tag":   rl.menuCompletePrevTag,
+		"accept-and-menu-complete": rl.acceptAndMenuComplete,
+		"vi-registers-complete":    rl.viRegistersComplete,
+		"menu-incremental-search":  rl.menuIncrementalSearch,
 	}
 }
 
@@ -32,79 +29,74 @@ func (rl *Shell) completionCommands() commands {
 // Commands ---------------------------------------------------------------------------
 //
 
+// Attempt completion on the current word.
+// Currently identitical to menu-complete.
 func (rl *Shell) completeWord() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
 	// This completion function should attempt to insert the first
 	// valid completion found, without printing the actual list.
+	if !rl.completer.IsActive() {
+		rl.startMenuComplete(rl.commandCompletion)
 
-	// switch rl.local {
-	// case menuselect, isearch:
-	// 	rl.menuComplete()
-	// default:
-	// 	// rl.
-	// 	// rl.startMenuComplete(rl.normalCompletions)
-	//
-	// 	// In autocomplete mode, we already have completions
-	// 	// printed, so we automatically move to the first comp.
-	// 	if rl.isAutoCompleting() && rl.local == menuselect {
-	// 		rl.menuComplete()
-	// 	}
-	// }
+		if rl.Config.GetBool("menu-complete-display-prefix") {
+			return
+		}
+	}
+
+	rl.completer.Select(1, 0)
+	rl.completer.SkipDisplay()
 }
 
+// List possible completions for the current word.
 func (rl *Shell) possibleCompletions() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
-	// switch rl.local {
-	// case menuselect, isearch:
-	// 	rl.resetVirtualComp(false)
-	// }
-	//
-	// rl.local = menuselect
-	//
-	// // Call the completer to produce
-	// // all possible completions.
-	// rl.normalCompletions()
-	//
-	// // Cancel completion mode if
-	// // we don't have any candidates.
-	// if rl.noCompletions() {
-	// 	rl.resetCompletion()
-	// 	return
-	// }
-
-	rl.completer.Cancel(false, false)
-	rl.keymaps.SetLocal(keymap.MenuSelect)
-	rl.completer.GenerateWith(rl.commandCompletion)
+	rl.startMenuComplete(rl.commandCompletion)
 }
 
-func (rl *Shell) insertCompletions() {}
+func (rl *Shell) insertCompletions() {
+	rl.History.Save()
 
+	// Generate all possible completions
+	if !rl.completer.IsActive() {
+		rl.startMenuComplete(rl.commandCompletion)
+	}
+
+	// Insert each match, cancel insertion with preserving
+	// the candidate just inserted in the line, for each.
+	for i := 0; i < rl.completer.Matches(); i++ {
+		rl.completer.Select(1, 0)
+		rl.completer.Cancel(false, false)
+	}
+
+	// Clear the completion menu.
+	rl.completer.Cancel(false, false)
+	rl.completer.ClearMenu(true)
+}
+
+// Like complete-word, except that menu completion is used.
 func (rl *Shell) menuComplete() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
 	// No completions are being printed yet, so simply generate the completions
 	// as if we just request them without immediately selecting a candidate.
 	if !rl.completer.IsActive() {
 		rl.startMenuComplete(rl.commandCompletion)
-	} else {
-		rl.completer.Select(1, 0)
+
+		// Immediately select only if not asked to display first.
+		if rl.Config.GetBool("menu-complete-display-prefix") {
+			return
+		}
 	}
 
-	// No completions are being printed yet, so simply generate the completions
-	// as if we just request them without immediately selecting a candidate.
-	// if rl.local != menuselect && rl.local != isearch && len(rl.histHint) == 0 {
-	// 	rl.startMenuComplete(rl.normalCompletions)
-	// }
-
-	// Some of the actions taken in the above switch might have exited
-	// completions, and if that is the case, we should not do anything.
-	// if rl.local != menuselect && rl.local != isearch && len(rl.histHint) == 0 {
-	// 	return
-	// }
+	rl.completer.Select(1, 0)
 }
 
+// Deletes the character under the cursor if not at the
+// beginning or end of the line (like delete-char).
+// If at the end of the line, behaves identically to
+// possible-completions.
 func (rl *Shell) deleteCharOrList() {
 	switch {
 	case rl.cursor.Pos() < rl.line.Len():
@@ -114,57 +106,24 @@ func (rl *Shell) deleteCharOrList() {
 	}
 }
 
-func (rl *Shell) expandOrComplete() {
-	// switch rl.local {
-	// case menuselect, isearch:
-	// 	rl.menuComplete()
-	// default:
-	// 	if rl.completer != nil {
-	// 		rl.startMenuComplete(rl.completer)
-	// 	} else {
-	// 		rl.startMenuComplete(rl.normalCompletions)
-	// 	}
-	//
-	// 	// In autocomplete mode, we already have completions
-	// 	// printed, so we automatically move to the first comp.
-	// 	if rl.isAutoCompleting() && rl.local == menuselect {
-	// 		rl.menuComplete()
-	// 	}
-	// }
-}
-
-func (rl *Shell) menuExpandOrComplete() {
-	// switch rl.local {
-	// case menuselect, isearch:
-	// 	rl.menuComplete()
-	// default:
-	// 	if rl.completer != nil {
-	// 		rl.startMenuComplete(rl.completer)
-	// 	} else {
-	// 		rl.startMenuComplete(rl.normalCompletions)
-	// 	}
-	//
-	// 	// In autocomplete mode, we already have completions
-	// 	// printed, so we automatically move to the first comp.
-	// 	if rl.isAutoCompleting() && rl.local == menuselect {
-	// 		rl.menuComplete()
-	// 	}
-	// }
-}
-
+// Identical to menu-complete, but moves backward through the
+// list of possible completions, as if menu-complete had been
+// given a negative argument.
 func (rl *Shell) menuCompleteBackward() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
 	// We don't do anything when not already completing.
 	if !rl.completer.IsActive() {
-		return
+		rl.startMenuComplete(rl.commandCompletion)
 	}
 
 	rl.completer.Select(-1, 0)
 }
 
+// In a menu completion, if there are several tags
+// of completions, go to the first result of the next tag.
 func (rl *Shell) menuCompleteNextTag() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
 	if !rl.completer.IsActive() {
 		return
@@ -173,8 +132,10 @@ func (rl *Shell) menuCompleteNextTag() {
 	rl.completer.SelectTag(true)
 }
 
+// In a menu completion, if there are several tags of
+// completions, go to the first result of the previous tag.
 func (rl *Shell) menuCompletePrevTag() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
 	if !rl.completer.IsActive() {
 		return
@@ -183,57 +144,44 @@ func (rl *Shell) menuCompletePrevTag() {
 	rl.completer.SelectTag(false)
 }
 
+// In a menu completion, insert the current completion
+// into the buffer, and advance to the next possible completion.
 func (rl *Shell) acceptAndMenuComplete() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
-	// // We don't do anything when not already completing.
-	// if rl.local != menuselect && rl.local != isearch {
-	// 	return
-	// }
-	//
-	// // Also return if no candidate
-	// if rl.currentCandidate() == "" {
-	// 	return
-	// }
-	//
-	// // First insert the current candidate
-	// rl.resetVirtualComp(false)
-	//
-	// // And cycle to the next one, without quiting our mode
-	// rl.updateSelector(1, 0)
-	// rl.updateVirtualComp()
+	// We don't do anything when not already completing.
+	if !rl.completer.IsActive() {
+		return
+	}
+
+	// Also return if no candidate
+	if !rl.completer.IsInserting() {
+		return
+	}
+
+	// First insert the current candidate.
+	rl.completer.Cancel(false, false)
+
+	// And cycle to the next one.
+	rl.completer.Select(1, 0)
 }
 
-func (rl *Shell) acceptCompletionOrLine() {
-	// switch rl.local {
-	// case menuselect, isearch:
-	// 	// If we have a completion, simply accept this candidate
-	// 	comp := rl.currentCandidate()
-	// 	if comp != "" {
-	// 		rl.resetVirtualComp(false)
-	// 		rl.resetCompletion()
-	//
-	// 		return
-	// 	}
-	//
-	// 	// Or accept the line.
-	// 	fallthrough
-	// default:
-	// 	rl.lineCarriageReturn()
-	// 	rl.accepted = true
-	// }
-}
-
+// Open a completion menu (similar to menu-complete) with all currently populated Vim registers.
 func (rl *Shell) viRegistersComplete() {
-	rl.keymaps.SetLocal(keymap.MenuSelect)
-	rl.undo.SkipSave()
-
-	registers := rl.buffers.Complete()
-	rl.completer.Generate(registers)
+	rl.History.SkipSave()
+	rl.startMenuComplete(rl.Buffers.Complete)
 }
 
+// In a menu completion (wether a candidate is selected or not), start incremental-search
+// (fuzzy search) on the results. Search backward incrementally for a specified string.
+// The search is case-insensitive if the search string does not have uppercase letters
+// and no numeric argument was given. The string may begin with ‘^’ to anchor the search
+// to the beginning of the line. A restricted set of editing functions is available in the
+// mini-buffer. Keys are looked up in the special isearch keymap, On each change in the
+// mini-buffer, any currently selected candidate is dropped from the line and the menu.
+// An interrupt signal, as defined by the stty setting, will stop the search and go back to the original line.
 func (rl *Shell) menuIncrementalSearch() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
 	if !rl.completer.IsActive() {
 		rl.completer.GenerateWith(rl.commandCompletion)
@@ -249,12 +197,13 @@ func (rl *Shell) menuIncrementalSearch() {
 // startMenuComplete generates a completion menu with completions
 // generated from a given completer, without selecting a candidate.
 func (rl *Shell) startMenuComplete(completer completion.Completer) {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
-	rl.keymaps.SetLocal(keymap.MenuSelect)
-	rl.completer.GenerateWith(rl.commandCompletion)
+	rl.Keymap.SetLocal(keymap.MenuSelect)
+	rl.completer.GenerateWith(completer)
 }
 
+// commandCompletion generates the completions for commands/args/flags.
 func (rl *Shell) commandCompletion() completion.Values {
 	if rl.Completer == nil {
 		return completion.Values{}
@@ -266,37 +215,45 @@ func (rl *Shell) commandCompletion() completion.Values {
 	return comps.convert()
 }
 
-func (rl *Shell) historyCompletion(forward, filterLine bool) {
-	switch rl.keymaps.Local() {
-	case keymap.MenuSelect, keymap.Isearch:
+// historyCompletion manages the various completion/isearch modes related
+// to history control. It can start the history completions, stop them, cycle
+// through sources if more than one, and adjust the completion/isearch behavior.
+func (rl *Shell) historyCompletion(forward, filterLine, substring bool) {
+	switch {
+	case rl.Keymap.Local() == keymap.MenuSelect || rl.Keymap.Local() == keymap.Isearch || rl.completer.AutoCompleting():
 		// If we are currently completing the last
 		// history source, cancel history completion.
-		if rl.histories.OnLastSource() {
-			rl.histories.Cycle(true)
+		if rl.History.OnLastSource() {
+			rl.History.Cycle(true)
 			rl.completer.ResetForce()
-			rl.hint.Reset()
+			rl.Hint.Reset()
 
 			return
 		}
 
 		// Else complete the next history source.
-		rl.histories.Cycle(true)
+		rl.History.Cycle(true)
 
 		fallthrough
 
 	default:
 		// Notify if we don't have history sources at all.
-		if rl.histories.Current() == nil {
-			rl.hint.Set(fmt.Sprintf("%s%s%s %s", color.Dim, color.FgRed, "No command history source", color.Reset))
+		if rl.History.Current() == nil {
+			rl.Hint.SetTemporary(fmt.Sprintf("%s%s%s %s", color.Dim, color.FgRed, "No command history source", color.Reset))
 			return
 		}
 
 		// Generate the completions with specified behavior.
 		completer := func() completion.Values {
-			return rl.histories.Complete(forward, filterLine)
+			return rl.History.Complete(forward, filterLine)
 		}
 
-		rl.completer.GenerateWith(completer)
-		rl.completer.IsearchStart(rl.histories.Name(), true)
+		if substring {
+			rl.completer.GenerateWith(completer)
+			rl.completer.IsearchStart(rl.History.Name(), true)
+		} else {
+			rl.startMenuComplete(completer)
+			rl.completer.AutocompleteForce()
+		}
 	}
 }
