@@ -3,18 +3,18 @@ package readline
 import (
 	"fmt"
 	"io"
-	"os"
-	"path/filepath"
 	"sort"
 	"strings"
 	"unicode"
 
 	"github.com/reeflective/readline/inputrc"
 	"github.com/reeflective/readline/internal/color"
+	"github.com/reeflective/readline/internal/completion"
 	"github.com/reeflective/readline/internal/editor"
 	"github.com/reeflective/readline/internal/keymap"
 	"github.com/reeflective/readline/internal/strutil"
 	"github.com/reeflective/readline/internal/term"
+	"github.com/rivo/uniseg"
 )
 
 // standardCommands returns all standard/emacs commands.
@@ -43,11 +43,11 @@ func (rl *Shell) standardCommands() commands {
 		"shell-backward-word":  rl.backwardShellWord,
 		"beginning-of-line":    rl.beginningOfLine,
 		"end-of-line":          rl.endOfLine,
-		"previous-screen-line": rl.upLine,   // up-line
-		"next-screen-line":     rl.downLine, // down-line
+		"previous-screen-line": rl.upLine,
+		"next-screen-line":     rl.downLine,
 		"clear-screen":         rl.clearScreen,
 		"clear-display":        rl.clearDisplay,
-		"redraw-current-line":  rl.display.Refresh,
+		"redraw-current-line":  rl.Display.Refresh,
 
 		// Changing text
 		"end-of-file":                  rl.endOfFile,
@@ -57,7 +57,7 @@ func (rl *Shell) standardCommands() commands {
 		"quoted-insert":                rl.quotedInsert,
 		"tab-insert":                   rl.tabInsert,
 		"self-insert":                  rl.selfInsert,
-		"bracketed-paste-begin":        rl.bracketedPasteBegin, // TODO: Finish and find how to do it.
+		"bracketed-paste-begin":        rl.bracketedPasteBegin,
 		"transpose-chars":              rl.transposeChars,
 		"transpose-words":              rl.transposeWords,
 		"shell-transpose-words":        rl.shellTransposeWords,
@@ -74,24 +74,24 @@ func (rl *Shell) standardCommands() commands {
 		"keyword-decrease": rl.keywordDecrease,
 
 		// Killing & yanking
-		"kill-line":                rl.killLine,
-		"backward-kill-line":       rl.backwardKillLine,
-		"unix-line-discard":        rl.backwardKillLine,
-		"kill-whole-line":          rl.killWholeLine,
-		"kill-word":                rl.killWord,
-		"backward-kill-word":       rl.backwardKillWord,
+		"kill-line":           rl.killLine,
+		"backward-kill-line":  rl.backwardKillLine,
+		"unix-line-discard":   rl.backwardKillLine,
+		"kill-whole-line":     rl.killWholeLine,
+		"kill-word":           rl.killWord,
+		"backward-kill-word":  rl.backwardKillWord,
+		"unix-word-rubout":    rl.backwardKillWord,
+		"kill-region":         rl.killRegion,
+		"copy-region-as-kill": rl.copyRegionAsKill,
+		"copy-backward-word":  rl.copyBackwardWord,
+		"copy-forward-word":   rl.copyForwardWord,
+		"yank":                rl.yank,
+		"yank-pop":            rl.yankPop,
+
+		"kill-buffer":              rl.killBuffer,
 		"shell-kill-word":          rl.shellKillWord,
 		"shell-backward-kill-word": rl.shellBackwardKillWord,
-		"unix-word-rubout":         rl.backwardKillWord,
-		"kill-region":              rl.killRegion,
-		"copy-region-as-kill":      rl.copyRegionAsKill,
-		"copy-backward-word":       rl.copyBackwardWord,
-		"copy-forward-word":        rl.copyForwardWord,
-		"yank":                     rl.yank,
-		"yank-pop":                 rl.yankPop,
-
-		"kill-buffer":          rl.killBuffer,
-		"copy-prev-shell-word": rl.copyPrevShellWord,
+		"copy-prev-shell-word":     rl.copyPrevShellWord,
 
 		// Numeric arguments
 		"digit-argument": rl.digitArgument,
@@ -102,6 +102,9 @@ func (rl *Shell) standardCommands() commands {
 		"call-last-kbd-macro":  rl.callLastKeyboardMacro,
 		"print-last-kbd-macro": rl.printLastKeyboardMacro,
 
+		"macro-toggle-record": rl.macroToggleRecord,
+		"macro-run":           rl.macroRun,
+
 		// Miscellaneous
 		"re-read-init-file":         rl.reReadInitFile,
 		"abort":                     rl.abort,
@@ -109,7 +112,7 @@ func (rl *Shell) standardCommands() commands {
 		"prefix-meta":               rl.prefixMeta,
 		"undo":                      rl.undoLast,
 		"revert-line":               rl.revertLine,
-		"set-mark":                  rl.setMark, // set-mark-command
+		"set-mark":                  rl.setMark,
 		"exchange-point-and-mark":   rl.exchangePointAndMark,
 		"character-search":          rl.characterSearch,
 		"character-search-backward": rl.characterSearchBackward,
@@ -121,7 +124,9 @@ func (rl *Shell) standardCommands() commands {
 		"edit-and-execute-command":  rl.editAndExecuteCommand,
 		"edit-command-line":         rl.editCommandLine,
 
-		"redo": rl.redo,
+		"redo":                rl.redo,
+		"select-keyword-next": rl.selectKeywordNext,
+		"select-keyword-prev": rl.selectKeywordPrev,
 	}
 
 	return widgets
@@ -131,41 +136,46 @@ func (rl *Shell) standardCommands() commands {
 // Modes ----------------------------------------------------------------
 //
 
+// When in vi command mode, this causes a switch to emacs editing mode.
 func (rl *Shell) emacsEditingMode() {
-	rl.keymaps.SetMain(keymap.Emacs)
+	rl.Keymap.SetMain(keymap.Emacs)
 }
 
 //
 // Movement -------------------------------------------------------------
 //
 
+// Move forward one character.
 func (rl *Shell) forwardChar() {
 	// Only exception where we actually don't forward a character.
-	if rl.opts.GetBool("history-autosuggest") && rl.cursor.Pos() == rl.line.Len()-1 {
+	if rl.Config.GetBool("history-autosuggest") && rl.cursor.Pos() == rl.line.Len()-1 {
 		rl.autosuggestAccept()
 		return
 	}
 
-	rl.undo.SkipSave()
-	vii := rl.iterations.Get()
+	rl.History.SkipSave()
+	vii := rl.Iterations.Get()
 
 	for i := 1; i <= vii; i++ {
 		rl.cursor.Inc()
 	}
 }
 
+// Move backward one character.
 func (rl *Shell) backwardChar() {
-	rl.undo.SkipSave()
-	vii := rl.iterations.Get()
+	rl.History.SkipSave()
+	vii := rl.Iterations.Get()
 
 	for i := 1; i <= vii; i++ {
 		rl.cursor.Dec()
 	}
 }
 
+// Move to the beginning of the next word. The editor’s idea
+// of a word is any sequence of alphanumeric characters.
 func (rl *Shell) forwardWord() {
-	rl.undo.SkipSave()
-	vii := rl.iterations.Get()
+	rl.History.SkipSave()
+	vii := rl.Iterations.Get()
 
 	for i := 1; i <= vii; i++ {
 		// When we have an autosuggested history and if we are at the end
@@ -177,18 +187,23 @@ func (rl *Shell) forwardWord() {
 	}
 }
 
+// Move to the beginning of the current or previousword. The editor’s
+// idea of a word is any sequence of alphanumeric characters.
 func (rl *Shell) backwardWord() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
-	vii := rl.iterations.Get()
+	vii := rl.Iterations.Get()
 	for i := 1; i <= vii; i++ {
 		backward := rl.line.Backward(rl.line.Tokenize, rl.cursor.Pos())
 		rl.cursor.Move(backward)
 	}
 }
 
+// Move forward to the beginning of the next word.
+// The editor's idea of a word is defined by classic sh-style word splitting:
+// any non-spaced sequence of characters, or a quoted sequence.
 func (rl *Shell) forwardShellWord() {
-	vii := rl.iterations.Get()
+	vii := rl.Iterations.Get()
 
 	for i := 1; i <= vii; i++ {
 		rl.selection.SelectAShellWord()
@@ -197,8 +212,11 @@ func (rl *Shell) forwardShellWord() {
 	}
 }
 
+// Move to the beginning of the current or previous word.
+// The editor's idea of a word is defined by classic sh-style word splitting:
+// any non-spaced sequence of characters, or a quoted sequence.
 func (rl *Shell) backwardShellWord() {
-	vii := rl.iterations.Get()
+	vii := rl.Iterations.Get()
 
 	for i := 1; i <= vii; i++ {
 		// First go the beginning of the blank word
@@ -212,79 +230,87 @@ func (rl *Shell) backwardShellWord() {
 	}
 }
 
+// Move to the beginning of the line. If already at the beginning
+// of the line, move to the beginning of the previous line, if any.
 func (rl *Shell) beginningOfLine() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
 	// Handle 0 as iteration to Vim.
-	if !rl.keymaps.IsEmacs() && rl.iterations.IsSet() {
-		rl.iterations.Add("0")
+	if !rl.Keymap.IsEmacs() && rl.Iterations.IsSet() {
+		rl.Iterations.Add("0")
 		return
 	}
 
 	rl.cursor.BeginningOfLine()
 }
 
+// Move to the end of the line. If already at the end
+// of the line, move to the end of the next line, if any.
 func (rl *Shell) endOfLine() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 	// If in Vim command mode, cursor
 	// will be brought back once later.
 	rl.cursor.EndOfLineAppend()
 }
 
+// Move up one line if the current buffer has more than one line.
 func (rl *Shell) upLine() {
-	lines := rl.iterations.Get()
+	lines := rl.Iterations.Get()
 	rl.cursor.LineMove(lines * -1)
 }
 
+// Move down one line if the current buffer has more than one line.
 func (rl *Shell) downLine() {
-	lines := rl.iterations.Get()
+	lines := rl.Iterations.Get()
 	rl.cursor.LineMove(lines)
 }
 
+// Clear the current screen and redisplay the prompt and input line.
+// This does not clear the terminal's output buffer.
 func (rl *Shell) clearScreen() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
 	fmt.Print(term.CursorTopLeft)
 	fmt.Print(term.ClearScreen)
 
-	rl.prompt.PrimaryPrint()
-	rl.display.CursorToPos()
+	rl.Display.PrintPrimaryPrompt()
 }
 
+// Clear the current screen and redisplay the prompt and input line.
+// This does clear the terminal's output buffer.
 func (rl *Shell) clearDisplay() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
 	fmt.Print(term.CursorTopLeft)
 	fmt.Print(term.ClearDisplay)
 
-	rl.prompt.PrimaryPrint()
-	rl.display.CursorToPos()
+	rl.Display.PrintPrimaryPrompt()
 }
 
 //
 // Changing Text --------------------------------------------------------
 //
 
+// The character indicating end-of-file as set, for example,
+// by “stty”.  If this character is read when there are no
+// characters on the line, and point is at the beginning of
+// the line, readline interprets it as the end of input and
+// returns EOF.
 func (rl *Shell) endOfFile() {
 	switch rl.line.Len() {
 	case 0:
-		rl.display.AcceptLine()
-		rl.histories.Accept(false, false, io.EOF)
+		rl.Display.AcceptLine()
+		rl.History.Accept(false, false, io.EOF)
 	default:
 		rl.deleteChar()
 	}
 }
 
+// Delete the character under the cursor.
 func (rl *Shell) deleteChar() {
-	// Extract from bash documentation of readline:
-	// Delete the character at point.  If this function is bound
-	// to the same character as the tty EOF character, as C-d
-	//
-	// TODO: We should match the same behavior here.
+	rl.History.Save()
 
-	rl.undo.Save()
-
-	vii := rl.iterations.Get()
+	vii := rl.Iterations.Get()
 
 	// Delete the chars in the line anyway
 	for i := 1; i <= vii; i++ {
@@ -292,38 +318,39 @@ func (rl *Shell) deleteChar() {
 	}
 }
 
+// Delete the character behind the cursor.
+// If the character to delete is a matching character
+// (quote/brackets/braces,etc) and that the character
+// under the cursor is its matching counterpart, this
+// character will also be deleted.
 func (rl *Shell) backwardDeleteChar() {
-	if rl.keymaps.Main() == keymap.ViIns {
-		rl.undo.SkipSave()
+	if rl.Keymap.Main() == keymap.ViInsert {
+		rl.History.SkipSave()
 	} else {
-		rl.undo.Save()
+		rl.History.Save()
 	}
 
-	rl.completer.Update()
+	// We might currently have a selected candidate inserted,
+	// and thus we should accept it as part of the real input
+	// line before cutting any character.
+	completion.UpdateInserted(rl.completer)
 
 	if rl.cursor.Pos() == 0 {
 		return
 	}
 
-	vii := rl.iterations.Get()
+	vii := rl.Iterations.Get()
 
 	switch vii {
 	case 1:
-		var toDelete rune
-		var isSurround, matcher bool
-
-		if rl.line.Len() > rl.cursor.Pos() {
-			toDelete = (*rl.line)[rl.cursor.Pos()-1]
-			isSurround = strutil.IsBracket(toDelete) || toDelete == '\'' || toDelete == '"'
-			matcher = strutil.IsSurround(toDelete, (*rl.line)[rl.cursor.Pos()])
+		// Handle removal of autopairs characters.
+		if rl.Config.GetBool("autopairs") {
+			completion.AutopairDelete(rl.line, rl.cursor)
 		}
 
+		// And then delete the character under cursor.
 		rl.cursor.Dec()
 		rl.line.CutRune(rl.cursor.Pos())
-
-		if isSurround && matcher {
-			rl.line.CutRune(rl.cursor.Pos())
-		}
 
 	default:
 		for i := 1; i <= vii; i++ {
@@ -333,6 +360,8 @@ func (rl *Shell) backwardDeleteChar() {
 	}
 }
 
+// Delete the character under the cursor, unless the cursor is at the
+// end of the line, in which case the character behind the cursor is deleted.
 func (rl *Shell) forwardBackwardDeleteChar() {
 	switch rl.cursor.Pos() {
 	case rl.line.Len():
@@ -342,67 +371,79 @@ func (rl *Shell) forwardBackwardDeleteChar() {
 	}
 }
 
+// Add the next character that you type to the line verbatim.
+// This is how to insert characters like C-q, for example.
 func (rl *Shell) quotedInsert() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 	rl.completer.TrimSuffix()
 
-	done := rl.keymaps.PendingCursor()
+	done := rl.Keymap.PendingCursor()
 	defer done()
 
-	keys, _ := rl.keys.ReadArgument()
+	key, _ := rl.Keys.ReadKey()
 
-	quoted := []rune{}
+	quoted, _ := strutil.Quote(key)
 
-	for _, key := range keys {
-		switch {
-		case inputrc.IsControl(key):
-			quoted = append(quoted, '^')
-			quoted = append(quoted, inputrc.Decontrol(key))
-		default:
-			quoted = append(quoted, key)
-		}
-	}
-
-	rl.line.Insert(rl.cursor.Pos(), quoted...)
-	rl.cursor.Move(len(quoted))
+	rl.cursor.InsertAt(quoted...)
 }
 
+// Insert a tab character.
 func (rl *Shell) tabInsert() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
-	// tab := fmt.Sprint("\t")
-	// rl.line.Insert(rl.cursor.Pos(), '\t')
-	// rl.cursor.Move(1)
+	rl.cursor.InsertAt('\t')
 }
 
+// Insert the character typed.
 func (rl *Shell) selfInsert() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
 	// Handle suffix-autoremoval for inserted completions.
 	rl.completer.TrimSuffix()
 
-	key, empty := rl.keys.Peek()
-	if empty {
-		return
+	key := rl.Keys.Caller()
+
+	// Handle autopair insertion (for the closer only)
+	searching, _, _ := rl.completer.NonIncrementallySearching()
+	isearch := rl.Keymap.Local() == keymap.Isearch
+
+	if !searching && !isearch && rl.Config.GetBool("autopairs") {
+		if jump := completion.AutopairInsertOrJump(key[0], rl.line, rl.cursor); jump {
+			return
+		}
 	}
 
-	// Insert the unescaped version of the key, and update cursor position.
-	unescaped := inputrc.Unescape(string(key))
-	rl.line.Insert(rl.cursor.Pos(), []rune(unescaped)...)
-	rl.cursor.Move(len(unescaped))
+	var quoted []rune
+	var length int
+
+	if rl.Config.GetBool("output-meta") && key[0] != inputrc.Esc {
+		quoted = append(quoted, key[0])
+		length = uniseg.StringWidth(string(quoted))
+	} else {
+		quoted, length = strutil.Quote(key[0])
+	}
+
+	rl.cursor.InsertAt(quoted...)
+	rl.cursor.Move(-1 * len(quoted))
+	rl.cursor.Move(length)
 }
 
 func (rl *Shell) bracketedPasteBegin() {
-	fmt.Println("Keys:")
-	keys, _ := rl.keys.PeekAll()
-	fmt.Println(string(keys))
+	// keys, _ := rl.Keys.PeekAllBytes()
+	// fmt.Println(string(keys))
 }
 
+// Drag the character before point forward over the character
+// at point, moving point forward as well.  If point is at the
+// end of the line, then this transposes the two characters
+// before point.  Negative arguments have no effect.
 func (rl *Shell) transposeChars() {
 	if rl.cursor.Pos() < 2 || rl.line.Len() < 2 {
-		rl.undo.SkipSave()
+		rl.History.SkipSave()
 		return
 	}
+
+	rl.History.Save()
 
 	switch {
 	case rl.cursor.Pos() == rl.line.Len():
@@ -411,15 +452,20 @@ func (rl *Shell) transposeChars() {
 		(*rl.line)[rl.cursor.Pos()-2] = last
 		(*rl.line)[rl.cursor.Pos()-1] = blast
 	default:
-		last := (*rl.line)[rl.cursor.Pos()]
+		last := rl.cursor.Char()
 		blast := (*rl.line)[rl.cursor.Pos()-1]
 		(*rl.line)[rl.cursor.Pos()-1] = last
-		(*rl.line)[rl.cursor.Pos()] = blast
+		rl.cursor.ReplaceWith(blast)
 	}
 }
 
+// Drag the word before point past the word after point,
+// moving point over that word as well.  If point is at the
+// end of the line, this transposes the last two words on the
+// line. If a numeric argument is given, the word to transpose
+// is chosen backward.
 func (rl *Shell) transposeWords() {
-	rl.undo.Save()
+	rl.History.Save()
 
 	startPos := rl.cursor.Pos()
 	rl.cursor.ToFirstNonSpace(true)
@@ -427,19 +473,22 @@ func (rl *Shell) transposeWords() {
 
 	// Save the current word and move the cursor to its beginning
 	rl.viSelectInWord()
+	rl.selection.Visual(false)
 	toTranspose, tbpos, tepos, _ := rl.selection.Pop()
 
 	// Then move some number of words.
 	// Either use words backward (if we are at end of line) or forward.
 	rl.cursor.Set(tbpos)
-	if tepos == rl.line.Len()-1 || tepos == rl.line.Len() || rl.iterations.IsSet() {
+
+	if tepos >= rl.line.Len()-1 || rl.Iterations.IsSet() {
 		rl.backwardWord()
 	} else {
-		rl.forwardWord()
+		rl.viForwardWord()
 	}
 
 	// Save the word to transpose with
 	rl.viSelectInWord()
+	rl.selection.Visual(false)
 	transposeWith, wbpos, wepos, _ := rl.selection.Pop()
 
 	// We might be on the first word of the line,
@@ -468,8 +517,12 @@ func (rl *Shell) transposeWords() {
 	rl.cursor.Set(tepos)
 }
 
+// Drag the shell word before point past the shell word after point,
+// moving point over that shell word as well. If point is at the
+// end of the line, this transposes the last two words on the line.
+// If a numeric argument is given, the word to transpose is chosen backward.
 func (rl *Shell) shellTransposeWords() {
-	rl.undo.Save()
+	rl.History.Save()
 
 	startPos := rl.cursor.Pos()
 
@@ -504,8 +557,10 @@ func (rl *Shell) shellTransposeWords() {
 	rl.cursor.Set(tepos)
 }
 
+// Lowercase the current (or following) word. With a negative argument,
+// lowercase the previous word, but do not move point.
 func (rl *Shell) downCaseWord() {
-	rl.undo.Save()
+	rl.History.Save()
 
 	startPos := rl.cursor.Pos()
 
@@ -522,8 +577,10 @@ func (rl *Shell) downCaseWord() {
 	rl.cursor.Set(startPos)
 }
 
+// Uppercase the current (or following) word.  With a negative argument,
+// uppercase the previous word, but do not move point.
 func (rl *Shell) upCaseWord() {
-	rl.undo.Save()
+	rl.History.Save()
 
 	startPos := rl.cursor.Pos()
 
@@ -540,8 +597,14 @@ func (rl *Shell) upCaseWord() {
 	rl.cursor.Set(startPos)
 }
 
+// Capitalize the current (or following) word.  With a negative argument,
+// capitalize the previous word, but do not move point.
 func (rl *Shell) capitalizeWord() {
-	rl.undo.Save()
+	if rl.line.Len() == 0 {
+		return
+	}
+
+	rl.History.Save()
 
 	startPos := rl.cursor.Pos()
 
@@ -549,53 +612,74 @@ func (rl *Shell) capitalizeWord() {
 	backward := rl.line.Backward(rl.line.Tokenize, rl.cursor.Pos())
 	rl.cursor.Move(backward)
 
-	letter := (*rl.line)[rl.cursor.Pos()]
-	upper := unicode.ToUpper(letter)
-	(*rl.line)[rl.cursor.Pos()] = upper
+	letter := rl.cursor.Char()
+	rl.cursor.ReplaceWith(unicode.ToUpper(letter))
 	rl.cursor.Set(startPos)
 }
 
+// Toggle overwrite mode. In overwrite mode, characters bound to
+// self-insert replace the text at point rather than pushing the
+// text to the right.  Characters bound to backward-delete-char
+// replace the character before point with a space.
 func (rl *Shell) overwriteMode() {
 	// We store the current line as an undo item first, but will not
 	// store any intermediate changes (in the loop below) as undo items.
-	rl.undo.Save()
+	rl.History.Save()
 
-	done := rl.keymaps.PendingCursor()
+	done := rl.Keymap.PendingCursor()
 	defer done()
+
+	// All replaced characters are stored, to be used with backspace
+	cache := make([]rune, 0)
+
+	// Don't use the delete cache past the end of the line
+	lineStart := rl.line.Len()
 
 	// The replace mode is quite special in that it does escape back
 	// to the main readline loop: it keeps reading characters and inserts
 	// them as long as the escape key is not pressed.
 	for {
-		// Read a new key
-		keys, esc := rl.keys.ReadArgument()
-		if esc {
-			return
+		// We read a character to use first.
+		key, isAbort := rl.Keys.ReadKey()
+		if isAbort {
+			break
 		}
 
-		key := keys[0]
-
 		// If the key is a backspace, we go back one character
-		if key == inputrc.Backspace {
-			rl.backwardDeleteChar()
+		if string(key) == inputrc.Unescape(string(`\C-?`)) {
+			if rl.cursor.Pos() > lineStart {
+				rl.backwardDeleteChar()
+			} else if rl.cursor.Pos() > 0 {
+				rl.cursor.Dec()
+			}
+
+			// And recover the last replaced character
+			if len(cache) > 0 && rl.cursor.Pos() < lineStart {
+				key = cache[len(cache)-1]
+				cache = cache[:len(cache)-1]
+
+				rl.cursor.ReplaceWith(key)
+			}
 		} else {
 			// If the cursor is at the end of the line,
 			// we insert the character instead of replacing.
-			if rl.cursor.Pos() == rl.line.Len() {
-				rl.line.Insert(rl.cursor.Pos(), key)
+			if rl.line.Len() == rl.cursor.Pos() {
+				rl.cursor.InsertAt(key)
 			} else {
-				(*rl.line)[rl.cursor.Pos()] = key
+				cache = append(cache, rl.cursor.Char())
+				rl.cursor.ReplaceWith(key)
+				rl.cursor.Inc()
 			}
-
-			rl.cursor.Inc()
 		}
 
-		rl.display.Refresh()
+		// Update the line
+		rl.Display.Refresh()
 	}
 }
 
+// Delete all spaces and tabs around point.
 func (rl *Shell) deleteHorizontalWhitespace() {
-	rl.undo.Save()
+	rl.History.Save()
 
 	startPos := rl.cursor.Pos()
 
@@ -604,21 +688,22 @@ func (rl *Shell) deleteHorizontalWhitespace() {
 	if rl.cursor.Pos() != startPos {
 		rl.cursor.Inc()
 	}
-	bpos := rl.cursor.Pos()
 
+	bpos := rl.cursor.Pos()
 	rl.cursor.ToFirstNonSpace(true)
 
 	if rl.cursor.Pos() != startPos {
 		rl.cursor.Dec()
 	}
-	epos := rl.cursor.Pos()
 
+	epos := rl.cursor.Pos()
 	rl.line.Cut(bpos, epos)
 	rl.cursor.Set(bpos)
 }
 
+// Delete the current word from the cursor point up to the end of it.
 func (rl *Shell) deleteWord() {
-	rl.undo.Save()
+	rl.History.Save()
 
 	rl.selection.Mark(rl.cursor.Pos())
 	forward := rl.line.ForwardEnd(rl.line.Tokenize, rl.cursor.Pos())
@@ -627,19 +712,19 @@ func (rl *Shell) deleteWord() {
 	rl.selection.Cut()
 }
 
+// Quote the region from the cursor to the mark.
 func (rl *Shell) quoteRegion() {
-	rl.undo.Save()
+	rl.History.Save()
 
 	rl.selection.Surround('\'', '\'')
 	rl.cursor.Inc()
 }
 
+// Quote the entire line buffer.
 func (rl *Shell) quoteLine() {
 	if rl.line.Len() == 0 {
 		return
 	}
-
-	rl.line.Insert(0, '\'')
 
 	for pos, char := range *rl.line {
 		if char == '\n' {
@@ -651,36 +736,36 @@ func (rl *Shell) quoteLine() {
 		}
 	}
 
+	rl.line.Insert(0, '\'')
 	rl.line.Insert(rl.line.Len(), '\'')
 }
 
+// Modifies the current word under the cursor, increasing it.
+// The following word types can be incremented/decremented:
+//
+//	Booleans: true|false, t|f, on|off, yes|no, y|n.
+//	Operators: &&|||, ++|--, ==|!=, ===| !==, +| -, -| *, *| /, /| +, and| or.
+//	Hex digits 0xDe => 0xdf, 0xdE => 0xDF, 0xde0 => 0xddf, 0xffffffffffffffff => 0x0000000000000000.
+//	Binary digits: 0b1 => 0b10, 0B0 => 0B1, etc.
+//	Integers.
 func (rl *Shell) keywordIncrease() {
-	rl.undo.Save()
+	rl.History.Save()
 	rl.keywordSwitch(true)
 }
 
+// Modifies the current word under the cursor, decreasing it.
+// The following word types can be incremented/decremented:
+//
+//	Booleans: true|false, t|f, on|off, yes|no, y|n.
+//	Operators: &&|||, ++|--, ==|!=, ===| !==, +| -, -| *, *| /, /| +, and| or.
+//	Hex digits 0xDe => 0xdf, 0xdE => 0xDF, 0xde0 => 0xddf, 0xffffffffffffffff => 0x0000000000000000.
+//	Binary digits: 0b1 => 0b10, 0B0 => 0B1, etc.
+//	Integers.
 func (rl *Shell) keywordDecrease() {
-	rl.undo.Save()
+	rl.History.Save()
 	rl.keywordSwitch(false)
 }
 
-// Cursor position cases:
-//
-// 1. Cursor on symbol:
-// 2+2   => +
-// 2-2   => -
-// 2 + 2 => +
-// 2 +2  => +2
-// 2 -2  => -2
-// 2 -a  => -a
-//
-// 2. Cursor on number or alpha:
-// 2+2   => +2
-// 2-2   => -2
-// 2 + 2 => 2
-// 2 +2  => +2
-// 2 -2  => -2
-// 2 -a  => -a.
 func (rl *Shell) keywordSwitch(increase bool) {
 	cpos := strutil.AdjustNumberOperatorPos(rl.cursor.Pos(), *rl.line)
 
@@ -699,7 +784,7 @@ func (rl *Shell) keywordSwitch(increase bool) {
 	// For each of the keyword handlers, run it, which returns
 	// false/none if didn't operate, then continue to next handler.
 	for _, switcher := range strutil.KeywordSwitchers() {
-		vii := rl.iterations.Get()
+		vii := rl.Iterations.Get()
 
 		changed, word, obpos, oepos := switcher(selection, increase, vii)
 		if !changed {
@@ -732,50 +817,71 @@ func (rl *Shell) keywordSwitch(increase bool) {
 // Killing & Yanking ----------------------------------------------------------
 //
 
+// Kill from the cursor to the end of the line. If already
+// on the end of the line, kill the newline character.
 func (rl *Shell) killLine() {
-	rl.iterations.Reset()
-	rl.undo.Save()
+	rl.Iterations.Reset()
+	rl.History.Save()
 
-	cut := []rune(*rl.line)[rl.cursor.Pos():]
-	rl.buffers.Write(cut...)
+	if rl.line.Len() == 0 {
+		return
+	}
 
-	rl.line.Cut(rl.cursor.Pos(), rl.line.Len())
+	cpos := rl.cursor.Pos()
+	rl.cursor.EndOfLineAppend()
+
+	rl.selection.MarkRange(cpos, rl.cursor.Pos())
+	text := rl.selection.Cut()
+
+	rl.Buffers.Write([]rune(text)...)
+	rl.cursor.Set(cpos)
 }
 
+// Kill backward to the beginning of the line.
 func (rl *Shell) backwardKillLine() {
-	rl.iterations.Reset()
-	rl.undo.Save()
+	rl.Iterations.Reset()
+	rl.History.Save()
 
-	cut := []rune(*rl.line)[:rl.cursor.Pos()]
-	rl.buffers.Write(cut...)
+	if rl.line.Len() == 0 {
+		return
+	}
 
-	rl.line.Cut(0, rl.cursor.Pos())
+	cpos := rl.cursor.Pos()
+	rl.cursor.BeginningOfLine()
+
+	rl.selection.MarkRange(rl.cursor.Pos(), cpos)
+	text := rl.selection.Cut()
+
+	rl.Buffers.Write([]rune(text)...)
 }
 
+// Kill all characters on the current line, no matter where point is.
 func (rl *Shell) killWholeLine() {
-	rl.undo.Save()
+	rl.History.Save()
 
 	if rl.line.Len() == 0 {
 		return
 	}
 
-	rl.buffers.Write(*rl.line...)
+	rl.Buffers.Write(*rl.line...)
 	rl.line.Cut(0, rl.line.Len())
 }
 
+// Kill the entire buffer.
 func (rl *Shell) killBuffer() {
-	rl.undo.Save()
+	rl.History.Save()
 
 	if rl.line.Len() == 0 {
 		return
 	}
 
-	rl.buffers.Write(*rl.line...)
+	rl.Buffers.Write(*rl.line...)
 	rl.line.Cut(0, rl.line.Len())
 }
 
+// Kill the current word from the cursor point up to the end of it.
 func (rl *Shell) killWord() {
-	rl.undo.Save()
+	rl.History.Save()
 
 	bpos := rl.cursor.Pos()
 
@@ -785,21 +891,97 @@ func (rl *Shell) killWord() {
 	epos := rl.cursor.Pos()
 
 	rl.selection.MarkRange(bpos, epos)
-	rl.buffers.Write([]rune(rl.selection.Cut())...)
+	rl.Buffers.Write([]rune(rl.selection.Cut())...)
 	rl.cursor.Set(bpos)
 }
 
+// Kill the word behind point. Word boundaries
+// are the same as those used by backward-word.
 func (rl *Shell) backwardKillWord() {
-	rl.undo.Save()
-	rl.undo.SkipSave()
+	rl.History.Save()
+	rl.History.SkipSave()
 
 	rl.selection.Mark(rl.cursor.Pos())
 	adjust := rl.line.Backward(rl.line.Tokenize, rl.cursor.Pos())
 	rl.cursor.Move(adjust)
 
-	rl.buffers.Write([]rune(rl.selection.Cut())...)
+	rl.Buffers.Write([]rune(rl.selection.Cut())...)
 }
 
+// Kill the text between the point and mark (saved cursor
+// position).  This text is referred to as the region.
+func (rl *Shell) killRegion() {
+	rl.History.Save()
+
+	if !rl.selection.Active() {
+		return
+	}
+
+	rl.Buffers.Write([]rune(rl.selection.Cut())...)
+}
+
+// Copy the text in the region to the kill buffer.
+func (rl *Shell) copyRegionAsKill() {
+	rl.History.SkipSave()
+
+	if !rl.selection.Active() {
+		return
+	}
+
+	rl.Buffers.Write([]rune(rl.selection.Text())...)
+	rl.selection.Reset()
+}
+
+// Copy the word before point to the kill buffer.
+// The word boundaries are the same as backward-word.
+func (rl *Shell) copyBackwardWord() {
+	rl.History.Save()
+
+	rl.selection.Mark(rl.cursor.Pos())
+	adjust := rl.line.Backward(rl.line.Tokenize, rl.cursor.Pos())
+	rl.cursor.Move(adjust)
+
+	rl.Buffers.Write([]rune(rl.selection.Text())...)
+	rl.selection.Reset()
+}
+
+// Copy the word following point to the kill buffer.
+// The word boundaries are the same as forward-word.
+func (rl *Shell) copyForwardWord() {
+	rl.History.Save()
+
+	rl.selection.Mark(rl.cursor.Pos())
+	adjust := rl.line.Forward(rl.line.Tokenize, rl.cursor.Pos())
+	rl.cursor.Move(adjust + 1)
+
+	rl.Buffers.Write([]rune(rl.selection.Text())...)
+	rl.selection.Reset()
+}
+
+// Yank the top of the kill ring into the buffer at point.
+func (rl *Shell) yank() {
+	buf := rl.Buffers.Active()
+
+	vii := rl.Iterations.Get()
+
+	for i := 1; i <= vii; i++ {
+		rl.cursor.InsertAt(buf...)
+	}
+}
+
+// Rotate the kill ring, and yank the new top.
+// Only works following yank or yank-pop.
+func (rl *Shell) yankPop() {
+	vii := rl.Iterations.Get()
+
+	for i := 1; i <= vii; i++ {
+		buf := rl.Buffers.Pop()
+		rl.cursor.InsertAt(buf...)
+	}
+}
+
+// Kill the shell word behind point. Word boundaries
+// are the same as those used by backward-word.
 func (rl *Shell) shellKillWord() {
 	startPos := rl.cursor.Pos()
 
@@ -809,133 +991,47 @@ func (rl *Shell) shellKillWord() {
 
 	_, epos := rl.selection.Pos()
 
-	rl.buffers.Write([]rune((*rl.line)[startPos:epos])...)
+	rl.Buffers.Write([]rune((*rl.line)[startPos:epos])...)
 	rl.line.Cut(startPos, epos)
 	rl.cursor.Set(startPos)
 
 	rl.selection.Reset()
 }
 
+// Kill the shell word behind point.
 func (rl *Shell) shellBackwardKillWord() {
 	startPos := rl.cursor.Pos()
+	if rl.line.Len() == 0 || startPos == 0 {
+		return
+	}
+
+	rl.History.Save()
 
 	// Always ignore the character under cursor.
 	rl.cursor.Dec()
 	rl.cursor.ToFirstNonSpace(false)
 
-	// Find the beginning of a correctly formatted shellword.
-	rl.viSelectAShellWord()
-	bpos, _ := rl.selection.Pos()
+	unclosed, bpos := strutil.GetQuotedWordStart((*rl.line)[:startPos])
 
-	// But don't include any of the leading spaces.
+	if !unclosed {
+		rl.viSelectAShellWord()
+		bpos, _ = rl.selection.Pos()
+	}
+
 	rl.cursor.Set(bpos)
 	rl.cursor.ToFirstNonSpace(true)
 	bpos = rl.cursor.Pos()
 
-	// Find any quotes backard, and settle on the outermost one.
-	outquote := -1
-
-	squote := rl.line.Find('\'', bpos, false)
-	dquote := rl.line.Find('"', bpos, false)
-
-	if squote != -1 {
-		outquote = squote
-	}
-	if dquote != -1 {
-		if squote != -1 && dquote < squote {
-			outquote = dquote
-		} else if squote == -1 {
-			outquote = dquote
-		}
-	}
-
-	// If any is found, try to find if it's matching another one.
-	if outquote != -1 {
-		sBpos, sEpos := rl.line.SurroundQuotes(true, outquote)
-		dBpos, dEpos := rl.line.SurroundQuotes(false, outquote)
-		mark, _ := strutil.AdjustSurroundQuotes(dBpos, dEpos, sBpos, sEpos)
-
-		// And if no matches have been found, only use the quote
-		// if its backward to our currently found shellword.
-		if mark == -1 && outquote < bpos {
-			bpos = outquote
-			rl.cursor.Set(bpos)
-		}
-	}
-
-	// Remove the selection.
-	rl.buffers.Write([]rune((*rl.line)[bpos:startPos])...)
+	rl.Buffers.Write([]rune((*rl.line)[bpos:startPos])...)
 	rl.line.Cut(bpos, startPos)
-
 	rl.selection.Reset()
 }
 
-func (rl *Shell) killRegion() {
-	rl.undo.Save()
-
-	if !rl.selection.Active() {
-		return
-	}
-
-	rl.buffers.Write([]rune(rl.selection.Cut())...)
-}
-
-func (rl *Shell) copyRegionAsKill() {
-	rl.undo.SkipSave()
-
-	if !rl.selection.Active() {
-		return
-	}
-
-	rl.buffers.Write([]rune(rl.selection.Text())...)
-	rl.selection.Reset()
-}
-
-func (rl *Shell) copyBackwardWord() {
-	rl.undo.Save()
-
-	rl.selection.Mark(rl.cursor.Pos())
-	adjust := rl.line.Backward(rl.line.Tokenize, rl.cursor.Pos())
-	rl.cursor.Move(adjust)
-
-	rl.buffers.Write([]rune(rl.selection.Text())...)
-	rl.selection.Reset()
-}
-
-func (rl *Shell) copyForwardWord() {
-	rl.undo.Save()
-
-	rl.selection.Mark(rl.cursor.Pos())
-	adjust := rl.line.Forward(rl.line.Tokenize, rl.cursor.Pos())
-	rl.cursor.Move(adjust + 1)
-
-	rl.buffers.Write([]rune(rl.selection.Text())...)
-	rl.selection.Reset()
-}
-
-func (rl *Shell) yank() {
-	buf := rl.buffers.Get(rune(0))
-
-	vii := rl.iterations.Get()
-
-	for i := 1; i <= vii; i++ {
-		rl.line.Insert(rl.cursor.Pos(), buf...)
-		rl.cursor.Move(len(buf))
-	}
-}
-
-func (rl *Shell) yankPop() {
-	vii := rl.iterations.Get()
-
-	for i := 1; i <= vii; i++ {
-		buf := rl.buffers.Pop()
-		rl.line.Insert(rl.cursor.Pos(), buf...)
-		rl.cursor.Move(len(buf))
-	}
-}
-
+// Like copy-prev-word, but the word is found by using shell parsing,
+// whereas copy-prev-word looks for blanks. This makes a difference
+// when the word is quoted and contains spaces.
 func (rl *Shell) copyPrevShellWord() {
-	rl.undo.Save()
+	rl.History.Save()
 
 	posInit := rl.cursor.Pos()
 
@@ -962,83 +1058,171 @@ func (rl *Shell) copyPrevShellWord() {
 // digitArgument is used both in Emacs and Vim modes,
 // but strips the Alt modifier used in Emacs mode.
 func (rl *Shell) digitArgument() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
-	keys, empty := rl.keys.PeekAll()
-	if empty {
-		return
-	}
-
-	rl.iterations.Add(string(keys))
+	keys := rl.Keys.Caller()
+	rl.Iterations.Add(string(keys))
 }
 
 //
 // Macros ----------------------------------------------------------------------
 //
 
+// Begin saving the characters typed into the current keyboard macro.
 func (rl *Shell) startKeyboardMacro() {
-	rl.macros.StartRecord()
+	rl.Macros.StartRecord(rune(0))
 }
 
+// Stop saving the characters typed into the current
+// keyboard macro and store the definition.
 func (rl *Shell) endKeyboardMacro() {
-	rl.macros.StopRecord()
+	keys := rl.Keys.Caller()
+	rl.Macros.StopRecord(keys)
 }
 
+// Re-execute the last keyboard macro defined, by making the
+// characters in the macro appear as if typed at the keyboard.
 func (rl *Shell) callLastKeyboardMacro() {
-	rl.macros.RunLastMacro()
+	rl.Macros.RunLastMacro()
 }
 
+// Print the last keyboard macro defined in a format suitable for the inputrc file.
 func (rl *Shell) printLastKeyboardMacro() {
-	rl.display.ClearHelpers()
+	rl.Display.ClearHelpers()
 
-	rl.macros.PrintLastMacro()
+	rl.Macros.PrintLastMacro()
 
-	rl.prompt.PrimaryPrint()
-	rl.display.Refresh()
+	rl.Prompt.PrimaryPrint()
+	rl.Display.Refresh()
+}
+
+// Either starts recording a macro (if not yet recording), or stops it.
+// If the command is about to start recording a macro, it will read an
+// additional argument key (must be a letter), to be used as the macro
+// "name", just like macro recording and use work in Vim.
+// This command thus works "Vim-style", and should probably be used only
+// when using Vim editing mode.
+func (rl *Shell) macroToggleRecord() {
+	if rl.Macros.Recording() {
+		keys := rl.Keys.Caller()
+		rl.Macros.StopRecord(keys)
+
+		return
+	}
+
+	done := rl.Keymap.PendingCursor()
+	defer done()
+
+	rl.Hint.SetTemporary(color.Dim + "REC (macro arg)")
+	rl.Display.Refresh()
+
+	key, isAbort := rl.Keys.ReadKey()
+	if isAbort {
+		return
+	}
+
+	rl.Macros.StartRecord(key)
+}
+
+// Reads a key from the keyboard, and runs the macro stored for this key identitier.
+// This mimics the Vim-style or running macros. If no macro is recorded for this key,
+// or if the key is invalid, nothing happens.
+func (rl *Shell) macroRun() {
+	done := rl.Keymap.PendingCursor()
+	defer done()
+
+	rl.Hint.SetTemporary(color.Dim + "Run (macro arg)")
+	rl.Display.Refresh()
+
+	key, isAbort := rl.Keys.ReadKey()
+	if isAbort {
+		return
+	}
+
+	rl.Macros.RunMacro(key)
 }
 
 //
 // Miscellaneous ---------------------------------------------------------------
 //
 
+// Read in the contents of the inputrc file, and incorporate
+// any bindings or variable assignments found there.
 func (rl *Shell) reReadInitFile() {
-	config := filepath.Join(os.Getenv("HOME"), ".inputrc")
+	main := rl.Keymap.Main()
 
-	err := inputrc.ParseFile(config, rl.opts)
+	err := rl.Keymap.ReloadConfig(rl.Opts...)
 	if err != nil {
-		rl.hint.Set(color.FgRed + "Inputrc reload error: " + err.Error())
-	} else {
-		rl.hint.Set(color.FgGreen + "Inputrc reloaded: " + config)
-	}
-}
-
-func (rl *Shell) abort() {
-	// Reset any visual selection and iterations.
-	rl.iterations.Reset()
-	rl.selection.Reset()
-
-	// Cancel completions and/or incremental search.
-	rl.hint.Reset()
-	rl.completer.ResetForce()
-}
-
-// func (rl *Instance) errorCtrlC() error {
-// 	rl.keys = ""
-//
-// 	// Or return the current command line
-// 	rl.clearHelpers()
-// 	moveCursorDown(rl.fullY - rl.posY)
-// 	fmt.Print("\r\n")
-//
-// 	return ErrCtrlC
-
-func (rl *Shell) doLowercaseVersion() {
-	rl.undo.SkipSave()
-
-	keys, empty := rl.keys.PeekAll()
-	if empty {
+		rl.Hint.SetTemporary(color.FgRed + "Inputrc reload error: " + err.Error())
 		return
 	}
+
+	// Reload keymap settings and cursor
+	newMain := rl.Keymap.Main()
+
+	if main != newMain {
+		switch newMain {
+		case keymap.Emacs, keymap.EmacsStandard, keymap.EmacsMeta, keymap.EmacsCtrlX:
+			rl.emacsEditingMode()
+		case keymap.Vi, keymap.ViCommand, keymap.ViMove:
+			rl.viCommandMode()
+		case keymap.ViInsert:
+			rl.viInsertMode()
+		}
+	}
+
+	// Notify successfully reloaded
+	rl.Hint.SetTemporary(color.FgGreen + "Inputrc reloaded")
+}
+
+// Abort the current editing command.
+func (rl *Shell) abort() {
+	// Reset any visual selection and iterations.
+	rl.Iterations.Reset()
+	rl.selection.Reset()
+
+	// Cancel active completion insertion and/or incremental search.
+	if rl.completer.AutoCompleting() || rl.completer.IsInserting() {
+		rl.Hint.Reset()
+		rl.completer.ResetForce()
+
+		return
+	}
+
+	// Cancel non-incremental search modes.
+	searching, _, _ := rl.completer.NonIncrementallySearching()
+	if searching {
+		rl.completer.NonIsearchStop()
+		return
+	}
+
+	// And only return to the caller if the abort was
+	// called by one of the builtin/config terminators.
+	// All others should generally be OS signals.
+	if rl.Keymap.InputIsTerminator() {
+		return
+	}
+
+	if rl.Config.GetBool("echo-control-characters") {
+		key := rl.Keys.Caller()
+		if key[0] == rune(inputrc.Unescape(`\C-C`)[0]) {
+			quoted, _ := strutil.Quote(key[0])
+			fmt.Print(string(quoted))
+		}
+	}
+
+	// If no line was active,
+	rl.Display.AcceptLine()
+	rl.History.Accept(false, false, ErrInterrupt)
+}
+
+// If the metafied character x is uppercase, run the command
+// that is bound to the corresponding metafied lowercase character.
+// The behavior is undefined if x is already lowercase.
+func (rl *Shell) doLowercaseVersion() {
+	rl.History.SkipSave()
+
+	keys := rl.Keys.Caller()
 
 	escapePrefix := false
 
@@ -1060,44 +1244,53 @@ func (rl *Shell) doLowercaseVersion() {
 	// Feed back the keys with meta prefix or encoding
 	if escapePrefix {
 		input := append([]rune{inputrc.Esc}, keys...)
-		rl.keys.Feed(false, true, input...)
+		rl.Keys.Feed(false, input...)
 	} else {
-		rl.keys.Feed(false, true, inputrc.Enmeta(keys[0]))
+		rl.Keys.Feed(false, inputrc.Enmeta(keys[0]))
 	}
 }
 
+// Metafy the next character typed.  ESC f is equivalent to Meta-f.
 func (rl *Shell) prefixMeta() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
-	done := rl.keymaps.PendingCursor()
+	done := rl.Keymap.PendingCursor()
 	defer done()
 
-	keys, isAbort := rl.keys.ReadArgument()
+	key, isAbort := rl.Keys.ReadKey()
 	if isAbort {
 		return
 	}
 
-	keys = append([]rune{inputrc.Esc}, keys...)
-
 	// And feed them back to be used on the next loop.
-	rl.keys.Feed(false, true, keys...)
+	keys := append([]rune{inputrc.Esc}, key)
+	rl.Keys.Feed(false, keys...)
 }
 
+// Incrementally undo the last text modification.
+// Note that when invoked from vi command mode, the full
+// prior change made in insert mode is reverted, the changes
+// having been merged when command mode was selected.
 func (rl *Shell) undoLast() {
-	rl.undo.Undo(rl.line, rl.cursor)
+	rl.History.Undo()
 }
 
+// Undo all changes made to this line.
+// This is like executing the undo command enough
+// times to return the line to its initial state.
 func (rl *Shell) revertLine() {
-	rl.undo.Revert(rl.line, rl.cursor)
+	rl.History.Revert()
 }
 
+// Set the mark to the point. If a numeric argument is
+// supplied, the mark is set to that position.
 func (rl *Shell) setMark() {
 	switch {
-	case rl.iterations.IsSet():
+	case rl.Iterations.IsSet():
 		rl.cursor.SetMark()
 	default:
 		cpos := rl.cursor.Pos()
-		mark := rl.iterations.Get()
+		mark := rl.Iterations.Get()
 
 		if mark > rl.line.Len()-1 {
 			return
@@ -1109,6 +1302,9 @@ func (rl *Shell) setMark() {
 	}
 }
 
+// Swap the point with the mark.  The current cursor position
+// is set to the saved position, and the old cursor position
+// is saved as the mark.
 func (rl *Shell) exchangePointAndMark() {
 	// Deactivate mark if out of bound
 	if rl.cursor.Mark() > rl.line.Len() {
@@ -1132,31 +1328,49 @@ func (rl *Shell) exchangePointAndMark() {
 	}
 }
 
+// A character is read and point is moved to the next
+// occurrence of that character.  A negative argument
+// searches for previous occurrences.
 func (rl *Shell) characterSearch() {
-	if rl.iterations.Get() < 0 {
+	if rl.Iterations.Get() < 0 {
 		rl.viFindChar(false, false)
 	} else {
 		rl.viFindChar(true, false)
 	}
 }
 
+// A character is read and point is moved to the previous
+// occurrence of that character.  A negative argument
+// searches for subsequent occurrences.
 func (rl *Shell) characterSearchBackward() {
-	if rl.iterations.Get() < 0 {
+	if rl.Iterations.Get() < 0 {
 		rl.viFindChar(true, false)
 	} else {
 		rl.viFindChar(false, false)
 	}
 }
 
+// Without a numeric argument, the value of the readline
+// comment-begin variable is inserted at the beginning of the
+// current line.  If a numeric argument is supplied, this
+// command acts as a toggle: if the characters at the
+// beginning of the line do not match the value of
+// comment-begin, the value is inserted, otherwise the
+// characters in comment-begin are deleted from the beginning
+// of the line.  In either case, the line is accepted as if a
+// newline had been typed.  The default value of
+// comment-begin makes the current line a shell comment.
+// If a numeric argument causes the comment character to be
+// removed, the line will be executed by the shell.
 func (rl *Shell) insertComment() {
-	comment := rl.opts.GetString("comment-begin")
+	comment := strings.Trim(rl.Config.GetString("comment-begin"), "\"")
 
 	switch {
-	case !rl.iterations.IsSet():
+	case !rl.Iterations.IsSet():
 		// Without numeric argument, insert comment at the beginning of the line.
 		cpos := rl.cursor.Pos()
 		rl.cursor.BeginningOfLine()
-		rl.line.Insert(rl.cursor.Pos(), []rune(comment)...)
+		rl.cursor.InsertAt([]rune(comment)...)
 		rl.cursor.Set(cpos)
 
 	default:
@@ -1184,62 +1398,74 @@ func (rl *Shell) insertComment() {
 	rl.acceptLineWith(false, false)
 }
 
+// Print all of the functions and their key bindings to the
+// readline output stream.  If a numeric argument is
+// supplied, the output is formatted in such a way that it
+// can be made part of an inputrc file.
 func (rl *Shell) dumpFunctions() {
-	rl.display.ClearHelpers()
+	rl.Display.ClearHelpers()
 	fmt.Println()
 
 	defer func() {
-		rl.prompt.PrimaryPrint()
-		rl.display.Refresh()
+		rl.Prompt.PrimaryPrint()
+		rl.Display.Refresh()
 	}()
 
-	inputrcFormat := rl.iterations.IsSet()
-	rl.keymaps.PrintBinds(inputrcFormat)
+	inputrcFormat := rl.Iterations.IsSet()
+	rl.Keymap.PrintBinds(inputrcFormat)
 }
 
+// Print all of the settable variables and their values to
+// the readline output stream.  If a numeric argument is
+// supplied, the output is formatted in such a way that it
+// can be made part of an inputrc file.
 func (rl *Shell) dumpVariables() {
-	rl.display.ClearHelpers()
+	rl.Display.ClearHelpers()
 	fmt.Println()
 
 	defer func() {
-		rl.prompt.PrimaryPrint()
-		rl.display.Refresh()
+		rl.Prompt.PrimaryPrint()
+		rl.Display.Refresh()
 	}()
 
 	// Get all variables and their values, alphabetically sorted.
 	var variables []string
 
-	for variable := range rl.opts.Vars {
+	for variable := range rl.Config.Vars {
 		variables = append(variables, variable)
 	}
 
 	sort.Strings(variables)
 
 	// Either print in inputrc format, or wordly one.
-	if rl.iterations.IsSet() {
+	if rl.Iterations.IsSet() {
 		for _, variable := range variables {
-			value := rl.opts.Vars[variable]
+			value := rl.Config.Vars[variable]
 			fmt.Printf("set %s %v\n", variable, value)
 		}
 	} else {
 		for _, variable := range variables {
-			value := rl.opts.Vars[variable]
+			value := rl.Config.Vars[variable]
 			fmt.Printf("%s is set to `%v'\n", variable, value)
 		}
 	}
 }
 
+// Print all of the readline key sequences bound to macros
+// and the strings they output.  If a numeric argument is
+// supplied, the output is formatted in such a way that it
+// can be made part of an inputrc file.
 func (rl *Shell) dumpMacros() {
-	rl.display.ClearHelpers()
+	rl.Display.ClearHelpers()
 	fmt.Println()
 
 	defer func() {
-		rl.prompt.PrimaryPrint()
-		rl.display.Refresh()
+		rl.Prompt.PrimaryPrint()
+		rl.Display.Refresh()
 	}()
 
 	// We print the macros bound to the current keymap only.
-	binds := rl.opts.Binds[string(rl.keymaps.Main())]
+	binds := rl.Config.Binds[string(rl.Keymap.Main())]
 	if len(binds) == 0 {
 		return
 	}
@@ -1254,7 +1480,7 @@ func (rl *Shell) dumpMacros() {
 
 	sort.Strings(macroBinds)
 
-	if rl.iterations.IsSet() {
+	if rl.Iterations.IsSet() {
 		for _, key := range macroBinds {
 			action := inputrc.Escape(binds[inputrc.Unescape(key)].Action)
 			fmt.Printf("\"%s\": \"%s\"\n", key, action)
@@ -1267,39 +1493,43 @@ func (rl *Shell) dumpMacros() {
 	}
 }
 
+// Invoke an editor on the current command line, and execute the result as shell commands.
+// Readline attempts to invoke $VISUAL, $EDITOR, and emacs as the editor, in that order.
 func (rl *Shell) editAndExecuteCommand() {
 	buffer := *rl.line
 
 	// Edit in editor
-	edited, err := editor.EditBuffer(buffer, "", "")
+	edited, err := editor.EditBuffer(buffer, "", "", rl.Keymap.IsEmacs())
 	if err != nil || (len(edited) == 0 && len(buffer) != 0) {
-		rl.undo.SkipSave()
+		rl.History.SkipSave()
 
 		errStr := strings.ReplaceAll(err.Error(), "\n", "")
 		changeHint := fmt.Sprintf(color.FgRed+"Editor error: %s", errStr)
-		rl.hint.Set(changeHint)
+		rl.Hint.SetTemporary(changeHint)
 
 		return
 	}
 
 	// Update our line and return it the caller.
 	rl.line.Set(edited...)
-	rl.display.AcceptLine()
-	rl.histories.Accept(false, false, nil)
+	rl.Display.AcceptLine()
+	rl.History.Accept(false, false, nil)
 }
 
+// Invoke an editor on the current command line.
+// Readline attempts to invoke $VISUAL, $EDITOR, and emacs as the editor, in that order.
 func (rl *Shell) editCommandLine() {
 	buffer := *rl.line
-	keymapCur := rl.keymaps.Main()
+	keymapCur := rl.Keymap.Main()
 
 	// Edit in editor
-	edited, err := editor.EditBuffer(buffer, "", "")
+	edited, err := editor.EditBuffer(buffer, "", "", rl.Keymap.IsEmacs())
 	if err != nil || (len(edited) == 0 && len(buffer) != 0) {
-		rl.undo.SkipSave()
+		rl.History.SkipSave()
 
 		errStr := strings.ReplaceAll(err.Error(), "\n", "")
 		changeHint := fmt.Sprintf(color.FgRed+"Editor error: %s", errStr)
-		rl.hint.Set(changeHint)
+		rl.Hint.SetTemporary(changeHint)
 
 		return
 	}
@@ -1314,6 +1544,53 @@ func (rl *Shell) editCommandLine() {
 	}
 }
 
+// Incrementally redo undone text modifications.
 func (rl *Shell) redo() {
-	rl.undo.Redo(rl.line, rl.cursor)
+	rl.History.Redo()
+}
+
+// Considers the blank word under cursor, and tries a series of regular expressions on it
+// to match various patterns: URL and their various subcomponents (host/path/params, etc).
+//
+// When one of the regular expressions succeeds, the match is visually selected,
+// otherwise nothing is selected (if selection was active, it will stay the same)
+//
+// When repeatedly calling this function while in visual selection mode, the shell will
+// cycle through either the current matcher's capturing subgroups (such as the parts of a URL),
+// or cycle through the next matcher (for instance, attempting to grap an IP after trying URL).
+func (rl *Shell) selectKeywordNext() {
+	rl.History.SkipSave()
+
+	// Always try to find a match within the blank word under cursor.
+	bpos, epos := rl.line.SelectBlankWord(rl.cursor.Pos())
+
+	// Run the regexp matchers.
+	_, epos, match := rl.selection.SelectKeyword(bpos, epos, true)
+	if !match {
+		return
+	}
+
+	// The matchers succeeded, we now have a selection active,
+	// but the cursor should be moved to the end of it.
+	rl.cursor.Set(epos)
+	rl.selection.Visual(false)
+}
+
+// Identical to select-keyword-prev, except that the matcher/subgroup cycling occurs backward.
+func (rl *Shell) selectKeywordPrev() {
+	rl.History.SkipSave()
+
+	// Always try to find a match within the blank word under cursor.
+	bpos, epos := rl.line.SelectBlankWord(rl.cursor.Pos())
+
+	// Run the regexp matchers.
+	_, epos, match := rl.selection.SelectKeyword(bpos, epos, false)
+	if !match {
+		return
+	}
+
+	// The matchers succeeded, we now have a selection active,
+	// but the cursor should be moved to the end of it.
+	rl.cursor.Set(epos)
+	rl.selection.Visual(false)
 }
