@@ -3,6 +3,7 @@ package readline
 import (
 	"strings"
 
+	"github.com/reeflective/readline/inputrc"
 	"github.com/reeflective/readline/internal/history"
 	"github.com/reeflective/readline/internal/strutil"
 )
@@ -30,20 +31,20 @@ var NewInMemoryHistory = history.NewInMemoryHistory
 // AddHistoryFromFile adds a command history source from a file path.
 // The name is used when using/searching the history source.
 func (rl *Shell) AddHistoryFromFile(name, filepath string) {
-	rl.histories.AddFromFile(name, filepath)
+	rl.History.AddFromFile(name, filepath)
 }
 
 // Add adds a source of history lines bound to a given name (printed above
 // this source when used). When only the default in-memory history is bound,
 // it's replaced with the provided source. Following ones are added to the list.
 func (rl *Shell) AddHistory(name string, source history.Source) {
-	rl.histories.Add(name, source)
+	rl.History.Add(name, source)
 }
 
 // Delete deletes one or more history source by name.
 // If no arguments are passed, all currently bound sources are removed.
 func (rl *Shell) DeleteHistory(sources ...string) {
-	rl.histories.Delete(sources...)
+	rl.History.Delete(sources...)
 }
 
 // historyCommands returns all history commands.
@@ -53,14 +54,14 @@ func (rl *Shell) DeleteHistory(sources ...string) {
 func (rl *Shell) historyCommands() commands {
 	widgets := map[string]func(){
 		"accept-line":                            rl.acceptLine,
-		"next-history":                           rl.downHistory, // down-history
-		"previous-history":                       rl.upHistory,   // up-history
+		"next-history":                           rl.downHistory,
+		"previous-history":                       rl.upHistory,
 		"beginning-of-history":                   rl.beginningOfHistory,
 		"end-of-history":                         rl.endOfHistory,
-		"operate-and-get-next":                   rl.acceptLineAndDownHistory, // accept-line-and-down-history
+		"operate-and-get-next":                   rl.acceptLineAndDownHistory,
 		"fetch-history":                          rl.fetchHistory,
-		"forward-search-history":                 rl.historyIncrementalSearchForward,  // history-incremental-search-forward
-		"reverse-search-history":                 rl.historyIncrementalSearchBackward, // history-incremental-search-backward
+		"forward-search-history":                 rl.forwardSearchHistory,
+		"reverse-search-history":                 rl.reverseSearchHistory,
 		"non-incremental-forward-search-history": rl.nonIncrementalForwardSearchHistory,
 		"non-incremental-reverse-search-history": rl.nonIncrementalReverseSearchHistory,
 		"history-search-forward":                 rl.historySearchForward,
@@ -68,27 +69,30 @@ func (rl *Shell) historyCommands() commands {
 		"history-substring-search-forward":       rl.historySubstringSearchForward,
 		"history-substring-search-backward":      rl.historySubstringSearchBackward,
 		"yank-last-arg":                          rl.yankLastArg,
+		"insert-last-argument":                   rl.yankLastArg,
 		"yank-nth-arg":                           rl.yankNthArg,
 		"magic-space":                            rl.magicSpace,
 
-		"accept-and-hold":                   rl.acceptAndHold,
-		"accept-and-infer-next-history":     rl.acceptAndInferNextHistory,
-		"down-line-or-history":              rl.downLineOrHistory,
-		"up-line-or-history":                rl.upLineOrHistory,
-		"up-line-or-search":                 rl.upLineOrSearch,
-		"down-line-or-search":               rl.downLineOrSearch,
-		"infer-next-history":                rl.inferNextHistory,
-		"beginning-of-buffer-or-history":    rl.beginningOfBufferOrHistory,
-		"beginning-history-search-forward":  rl.beginningHistorySearchForward,
-		"beginning-history-search-backward": rl.beginningHistorySearchBackward,
-		"end-of-buffer-or-history":          rl.endOfBufferOrHistory,
-		"beginning-of-line-hist":            rl.beginningOfLineHist,
-		"end-of-line-hist":                  rl.endOfLineHist,
-		"autosuggest-accept":                rl.autosuggestAccept,
-		"autosuggest-execute":               rl.autosuggestExecute,
-		"autosuggest-enable":                rl.autosuggestEnable,
-		"autosuggest-disable":               rl.autosuggestDisable,
-		"autosuggest-toggle":                rl.autosuggestToggle,
+		"accept-and-hold":                    rl.acceptAndHold,
+		"accept-and-infer-next-history":      rl.acceptAndInferNextHistory,
+		"down-line-or-history":               rl.downLineOrHistory,
+		"up-line-or-history":                 rl.upLineOrHistory,
+		"up-line-or-search":                  rl.upLineOrSearch,
+		"down-line-or-select":                rl.downLineOrSelect,
+		"infer-next-history":                 rl.inferNextHistory,
+		"beginning-of-buffer-or-history":     rl.beginningOfBufferOrHistory,
+		"end-of-buffer-or-history":           rl.endOfBufferOrHistory,
+		"beginning-of-line-hist":             rl.beginningOfLineHist,
+		"end-of-line-hist":                   rl.endOfLineHist,
+		"incremental-forward-search-history": rl.incrementalForwardSearchHistory,
+		"incremental-reverse-search-history": rl.incrementalReverseSearchHistory,
+		"history-source-next":                rl.historySourceNext,
+		"history-source-prev":                rl.historySourcePrev,
+		"autosuggest-accept":                 rl.autosuggestAccept,
+		"autosuggest-execute":                rl.autosuggestExecute,
+		"autosuggest-enable":                 rl.autosuggestEnable,
+		"autosuggest-disable":                rl.autosuggestDisable,
+		"autosuggest-toggle":                 rl.autosuggestToggle,
 	}
 
 	return widgets
@@ -98,89 +102,175 @@ func (rl *Shell) historyCommands() commands {
 // Standard ----------------------------------------------------------------
 //
 
+// Finish editing the buffer. Normally this causes the buffer to be executed as a shell command.
 func (rl *Shell) acceptLine() {
 	rl.acceptLineWith(false, false)
 }
 
+// Move to the next event in the history list.
 func (rl *Shell) downHistory() {
-	rl.undo.SkipSave()
-	rl.histories.Walk(-1)
+	rl.History.Save()
+	rl.History.Walk(-1)
 }
 
+// Move to the previous event in the history list.
 func (rl *Shell) upHistory() {
-	rl.undo.SkipSave()
-	rl.histories.Walk(1)
+	rl.History.Save()
+	rl.History.Walk(1)
 }
 
+// Move to the first event in the history list.
 func (rl *Shell) beginningOfHistory() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
-	history := rl.histories.Current()
+	history := rl.History.Current()
 	if history == nil {
 		return
 	}
 
-	rl.histories.Walk(history.Len())
+	rl.History.Walk(history.Len())
 }
 
+// Move to the last event in the history list.
 func (rl *Shell) endOfHistory() {
-	history := rl.histories.Current()
+	history := rl.History.Current()
 
 	if history == nil {
 		return
 	}
 
-	rl.histories.Walk(-history.Len() + 1)
+	rl.History.Walk(-history.Len() + 1)
 }
 
+// Execute the current line, and push the next history event on the buffer stack.
 func (rl *Shell) acceptLineAndDownHistory() {
-	// rl.inferLine = true // The next loop will retrieve a line by histPos.
-	// rl.acceptLine()
+	rl.acceptLineWith(true, false)
 }
 
-func (rl *Shell) fetchHistory() {}
-
-func (rl *Shell) historyIncrementalSearchForward() {
-	rl.undo.SkipSave()
-	rl.historyCompletion(true, false)
+// With a numeric argument, fetch that entry from the history
+// list and make it the current line.  Without an argument,
+// move back to the first entry in the history list.
+func (rl *Shell) fetchHistory() {
+	if rl.Iterations.IsSet() {
+		rl.History.Fetch(rl.Iterations.Get())
+	} else {
+		rl.History.Fetch(0)
+	}
 }
 
-func (rl *Shell) historyIncrementalSearchBackward() {
-	rl.undo.SkipSave()
-	rl.historyCompletion(false, false)
+// Search forward starting at the current line and moving `down' through
+// the history as necessary.  This is an incremental search, opening and
+// showing matching completions.
+func (rl *Shell) forwardSearchHistory() {
+	rl.History.SkipSave()
+
+	forward := true
+	filterLine := false
+	regexp := true
+
+	rl.historyCompletion(forward, filterLine, regexp)
 }
 
-func (rl *Shell) nonIncrementalForwardSearchHistory() {}
-func (rl *Shell) nonIncrementalReverseSearchHistory() {}
+// Search backward starting at the current line and moving `up' through
+// the history as necessary.  This is an incremental search, opening and
+// showing matching completions.
+func (rl *Shell) reverseSearchHistory() {
+	rl.History.SkipSave()
 
+	forward := false
+	filterLine := false
+	regexp := true
+
+	rl.historyCompletion(forward, filterLine, regexp)
+}
+
+// Search forward through the history starting at the current line
+// using a non-incremental search for a string supplied by the user.
+func (rl *Shell) nonIncrementalForwardSearchHistory() {
+	repeat := false
+	forward := true
+	regexp := true
+
+	rl.completer.NonIsearchStart(rl.History.Name(), repeat, forward, regexp)
+}
+
+// Search backward through the history starting at the current line
+// using a non-incremental search for a string supplied by the user.
+func (rl *Shell) nonIncrementalReverseSearchHistory() {
+	repeat := false
+	forward := false
+	regexp := true
+
+	rl.completer.NonIsearchStart(rl.History.Name(), repeat, forward, regexp)
+}
+
+// Search forward through the history for the string of characters
+// between the start of the current line and the point.  The search
+// string must match at the beginning of a history line.
+// This shows the completions in autocomplete mode.
 func (rl *Shell) historySearchForward() {
-	rl.undo.SkipSave()
+	rl.History.Save()
 
-	// And either roll to the next history source, or
-	// directly generate completions for the target history.
-	//
-	// Set the tab completion prefix as a filtering
-	// mechanism here: will be updated by the comps anyway.
-	// rl.historyCompletion(true, true)
+	usePos := true
+	forward := true
+	regexp := false
+
+	rl.History.InsertMatch(nil, nil, usePos, forward, regexp)
 }
 
+// Search backward through the history for the string of characters
+// between the start of the current line and the point.  The search
+// string must match at the beginning of a history line.
+// This shows the completions in autocomplete mode.
 func (rl *Shell) historySearchBackward() {
-	rl.undo.SkipSave()
+	rl.History.Save()
 
-	// And either roll to the next history source, or
-	// directly generate completions for the target history.
-	//
-	// Set the tab completion prefix as a filtering
-	// mechanism here: will be updated by the comps anyway.
-	// rl.historyCompletion(false, true)
+	usePos := true
+	forward := false
+	regexp := false
+
+	rl.History.InsertMatch(nil, nil, usePos, forward, regexp)
 }
 
-func (rl *Shell) historySubstringSearchForward()  {}
-func (rl *Shell) historySubstringSearchBackward() {}
+// Search forward through the history for the string of characters
+// between the start of the current line and the current cursor position.
+// The search string may match anywhere in a history line.
+// This is a non-incremental search.
+func (rl *Shell) historySubstringSearchForward() {
+	usePos := true
+	forward := true
+	regexp := true
 
+	rl.History.InsertMatch(rl.line, rl.cursor, usePos, forward, regexp)
+}
+
+// Search backward through the history for the string of characters
+// between the start of the current line and the current cursor position.
+// The search string may match anywhere in a history line.
+// This is a non-incremental search.
+func (rl *Shell) historySubstringSearchBackward() {
+	usePos := true
+	forward := false
+	regexp := true
+
+	rl.History.InsertMatch(rl.line, rl.cursor, usePos, forward, regexp)
+}
+
+// Insert the last argument to the previous command (the last
+// word of the previous history entry).  With a numeric
+// argument, behave exactly like yank-nth-arg.  Successive
+// calls to yank-last-arg move back through the history list,
+// inserting the last word (or the word specified by the
+// argument to the first call) of each line in turn.
+// Any numeric argument supplied to these successive calls
+// determines the direction to move through the history.
+// A negative argument switches the direction through the
+// history (back or forward).  The history expansion
+// facilities are used to extract the last argument, as if
+// the "!$" history expansion had been specified.
 func (rl *Shell) yankLastArg() {
 	// Get the last history line.
-	last := rl.histories.GetLast()
+	last := rl.History.GetLast()
 	if last == "" {
 		return
 	}
@@ -206,9 +296,16 @@ func (rl *Shell) yankLastArg() {
 	rl.cursor.Move(len(lastArg))
 }
 
+// Insert the first argument to the previous command (usually
+// the second word on the previous line) at point.  With an
+// argument n, insert the nth word from the previous command
+// (the words in the previous command begin with word 0).
+// A negative argument inserts the nth word from the end of the
+// previous command.  Once the argument n is computed, the argument
+// is extracted as if the "!n" history expansion had been specified.
 func (rl *Shell) yankNthArg() {
 	// Get the last history line.
-	last := rl.histories.GetLast()
+	last := rl.History.GetLast()
 	if last == "" {
 		return
 	}
@@ -222,12 +319,12 @@ func (rl *Shell) yankNthArg() {
 	var lastArg string
 
 	// Abort if the required position is out of bounds.
-	argNth := rl.iterations.Get()
+	argNth := rl.Iterations.Get()
 	if len(words) < argNth {
 		return
-	} else {
-		lastArg = words[argNth-1]
 	}
+
+	lastArg = words[argNth-1]
 
 	// Quote if required.
 	if strings.ContainsAny(lastArg, " \t") {
@@ -243,27 +340,72 @@ func (rl *Shell) yankNthArg() {
 	rl.cursor.Move(len(lastArg))
 }
 
-func (rl *Shell) magicSpace() {}
+// Perform history expansion on the current line and insert a space.
+// If the current blank word under cursor starts with an exclamation
+// mark, the word up to the cursor is matched as a prefix against
+// the history lines, and the first match is inserted in place of it.
+func (rl *Shell) magicSpace() {
+	cpos := rl.cursor.Pos()
+	lineLen := rl.line.Len()
+
+	// If no line, or the cursor is on a space, we can't perform.
+	if lineLen == 0 || (cpos == lineLen && (*rl.line)[cpos-1] == inputrc.Space) {
+		rl.selfInsert()
+		return
+	}
+
+	// Select the word around cursor.
+	rl.viSelectInBlankWord()
+	word, bpos, _, _ := rl.selection.Pop()
+	rl.cursor.Set(cpos)
+
+	// Fail if empty or not prefixed expandable.
+	if len(strings.TrimSpace(word)) == 0 {
+		rl.selfInsert()
+		return
+	}
+
+	if !strings.HasPrefix(word, "!") || word == "!" {
+		rl.selfInsert()
+		return
+	}
+
+	// Else, perform expansion on the remainder.
+	pattern := (*rl.line)[bpos+1:]
+	suggested := rl.History.Suggest(&pattern)
+
+	if string(suggested) == string(pattern) {
+		rl.selfInsert()
+		return
+	}
+
+	rl.History.Save()
+	rl.line.Cut(bpos, lineLen)
+	rl.line.Insert(bpos, suggested...)
+	rl.cursor.Set(bpos + suggested.Len())
+}
 
 //
 // Added -------------------------------------------------------------------
 //
 
+// Accept the current input line (execute it) and
+// keep it as the buffer on the next readline loop.
 func (rl *Shell) acceptAndHold() {
+	rl.acceptLineWith(false, true)
+}
+
+// Execute the contents of the buffer. Then search the history list for a line
+// matching the current one and push the event following onto the buffer stack.
+func (rl *Shell) acceptAndInferNextHistory() {
 	rl.acceptLineWith(true, false)
 }
 
-func (rl *Shell) acceptAndInferNextHistory() {
-	// rl.inferLine = true // The next loop will retrieve a line.
-	// rl.histPos = 0      // And will find it by trying to match one.
-	// rl.acceptLine()
-}
-
+// Move down a line in the buffer, or if already at the
+// bottom line, move to the next event in the history list.
 func (rl *Shell) downLineOrHistory() {
-	rl.undo.SkipSave()
-
-	times := rl.iterations.Get()
-	linesDown := rl.line.Lines() - rl.cursor.Line()
+	times := rl.Iterations.Get()
+	linesDown := rl.line.Lines() - rl.cursor.LinePos()
 
 	// If we can go down some lines out of
 	// the available iterations, use them.
@@ -273,15 +415,15 @@ func (rl *Shell) downLineOrHistory() {
 	}
 
 	if times > 0 {
-		rl.histories.Walk(times * -1)
+		rl.History.Walk(times * -1)
 	}
 }
 
+// Move up a line in the buffer, or if already at the top
+// line, move to the previous event in the history list.
 func (rl *Shell) upLineOrHistory() {
-	rl.undo.SkipSave()
-
-	times := rl.iterations.Get()
-	linesUp := rl.cursor.Line()
+	times := rl.Iterations.Get()
+	linesUp := rl.cursor.LinePos()
 
 	// If we can go down some lines out of
 	// the available iterations, use them.
@@ -291,89 +433,153 @@ func (rl *Shell) upLineOrHistory() {
 	}
 
 	if times > 0 {
-		rl.histories.Walk(times)
+		rl.History.Walk(times)
 	}
 }
 
+// If the cursor is on the first line of the buffer, start an incremental
+// search backward on the history lines. Otherwise, move up a line in the buffer.
 func (rl *Shell) upLineOrSearch() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
+
 	switch {
-	case rl.cursor.Line() > 0:
+	case rl.cursor.LinePos() > 0:
 		rl.cursor.LineMove(-1)
 	default:
 		rl.historySearchBackward()
 	}
 }
 
-func (rl *Shell) downLineOrSearch() {
-	rl.undo.SkipSave()
+// If the cursor is on the last line of the buffer, start an incremental
+// search forward on the history lines. Otherwise, move up a line in the buffer.
+func (rl *Shell) downLineOrSelect() {
+	rl.History.SkipSave()
+
 	switch {
-	case rl.cursor.Line() < rl.line.Lines():
+	case rl.cursor.LinePos() < rl.line.Lines():
 		rl.cursor.LineMove(1)
 	default:
-		rl.historySearchForward()
+		rl.menuComplete()
 	}
 }
 
+// Attempt to find a line in history matching the current line buffer as a prefix,
+// and if one is found, fetch the next history event and make it the current buffer.
 func (rl *Shell) inferNextHistory() {
-	rl.undo.SkipSave()
-	rl.histories.InferNext()
+	rl.History.SkipSave()
+	rl.History.InferNext()
 }
 
+// If the cursor is not at the beginning of the buffer, go to it.
+// Otherwise, go to the beginning of history.
 func (rl *Shell) beginningOfBufferOrHistory() {
-	rl.undo.SkipSave()
-
 	if rl.cursor.Pos() > 0 {
+		rl.History.SkipSave()
 		rl.cursor.Set(0)
+
 		return
 	}
 
+	rl.History.Save()
 	rl.beginningOfHistory()
 }
 
+// If the cursor is not at the end of the buffer, go to it.
+// Otherwise, go to the end of history.
 func (rl *Shell) endOfBufferOrHistory() {
-	rl.undo.SkipSave()
-
 	if rl.cursor.Pos() < rl.line.Len()-1 {
+		rl.History.SkipSave()
 		rl.cursor.Set(rl.line.Len())
+
 		return
 	}
 
+	rl.History.Save()
 	rl.endOfHistory()
 }
 
+// Go to the beginning of the current line, if the cursor is not yet.
+// If at the beginning of the line, attempt to move one line up.
+// If at the beginning of the buffer, move up one history line.
 func (rl *Shell) beginningOfLineHist() {
-	rl.undo.SkipSave()
-
 	switch {
-	// case rl.pos <= 0:
-	// 	rl.beginningOfLine()
+	case rl.cursor.Pos() > 0:
+		rl.History.SkipSave()
+
+		if rl.cursor.AtBeginningOfLine() {
+			rl.cursor.Dec()
+		}
+
+		rl.beginningOfLine()
 	default:
-		rl.histories.Walk(1)
+		rl.History.Save()
+		rl.History.Walk(1)
 	}
 }
 
+// Go to the end of the current line, if the cursor is not yet.
+// If at the end of the line, attempt to move one line down.
+// If at the end of the buffer, move up one history line.
 func (rl *Shell) endOfLineHist() {
-	rl.undo.SkipSave()
+	rl.History.SkipSave()
 
 	switch {
-	// case rl.cursor.Pos() < len(rl.line)-1:
-	// 	rl.endOfLine()
+	case rl.cursor.Pos() < rl.line.Len()-1:
+		rl.History.SkipSave()
+
+		if rl.cursor.AtEndOfLine() {
+			rl.cursor.Inc()
+		}
+
+		rl.endOfLine()
+
 	default:
-		rl.histories.Walk(-1)
+		rl.History.Save()
+		rl.History.Walk(-1)
 	}
 }
 
-func (rl *Shell) beginningHistorySearchBackward() {
-	// rl.historySearchLine(false)
+// Start an forward history autocompletion mode, starting at the
+// current line and moving `down' through the history as necessary.
+func (rl *Shell) incrementalForwardSearchHistory() {
+	rl.History.SkipSave()
+
+	forward := true
+	filter := true
+	regexp := false
+
+	rl.historyCompletion(forward, filter, regexp)
 }
 
-func (rl *Shell) beginningHistorySearchForward() {
-	// rl.historySearchLine(true)
+// Start an backward history autocompletion mode, starting at the
+// current line and moving `down' through the history as necessary.
+func (rl *Shell) incrementalReverseSearchHistory() {
+	rl.History.SkipSave()
+
+	forward := false
+	filter := true
+	regexp := false
+
+	rl.historyCompletion(forward, filter, regexp)
 }
 
+// If more than one source of command history is bound to the shell,
+// cycle to the next one and use it for all history search operations,
+// movements across lines, their respective undo histories, etc.
+func (rl *Shell) historySourceNext() {
+	rl.History.Cycle(true)
+}
+
+// If more than one source of command history is bound to the shell,
+// cycle to the previous one and use it for all history search operations,
+// movements across lines, their respective undo histories, etc.
+func (rl *Shell) historySourcePrev() {
+	rl.History.Cycle(false)
+}
+
+// If a line is currently autoggested, make it the buffer.
 func (rl *Shell) autosuggestAccept() {
-	suggested := rl.histories.Suggest(rl.line)
+	suggested := rl.History.Suggest(rl.line)
 
 	if suggested.Len() <= rl.line.Len() {
 		return
@@ -383,8 +589,9 @@ func (rl *Shell) autosuggestAccept() {
 	rl.cursor.Set(len(suggested))
 }
 
+// If a line is currently autoggested, make it the buffer and execute it.
 func (rl *Shell) autosuggestExecute() {
-	suggested := rl.histories.Suggest(rl.line)
+	suggested := rl.History.Suggest(rl.line)
 
 	if suggested.Len() <= rl.line.Len() {
 		return
@@ -396,22 +603,29 @@ func (rl *Shell) autosuggestExecute() {
 	rl.acceptLine()
 }
 
+// Toggle line history autoggestions on/off.
 func (rl *Shell) autosuggestToggle() {
-	if rl.opts.GetBool("history-autosuggest") {
+	if rl.Config.GetBool("history-autosuggest") {
 		rl.autosuggestDisable()
 	} else {
 		rl.autosuggestEnable()
 	}
 }
 
+// Enable history line autoggestions.
+// When enabled and if a line is suggested, forward-word commands, will
+// take the first word of the non-inserted part of this suggestion and
+// will insert it in the real input line.
+// The forward-char* commands, if at the end of the line, will accept it.
 func (rl *Shell) autosuggestEnable() {
-	rl.undo.SkipSave()
-	rl.opts.Vars["history-autosuggest"] = true
+	rl.History.SkipSave()
+	rl.Config.Vars["history-autosuggest"] = true
 }
 
+// Disable history line autoggestions.
 func (rl *Shell) autosuggestDisable() {
-	rl.undo.SkipSave()
-	rl.opts.Vars["history-autosuggest"] = false
+	rl.History.SkipSave()
+	rl.Config.Vars["history-autosuggest"] = false
 }
 
 //
@@ -423,20 +637,44 @@ func (rl *Shell) acceptLineWith(infer, hold bool) {
 	// we should cancel this mode so as to run the rest of this
 	// function on (with) the input line itself, not the minibuffer.
 	rl.completer.Reset()
+
+	// Non-incremental search modes are the only mode not cancelled
+	// by the completion engine. If it's active, match the line result
+	// and return without returning the line to the readline caller.
+	searching, forward, substring := rl.completer.NonIncrementallySearching()
+	if searching {
+		defer rl.completer.NonIsearchStop()
+
+		line, cursor, _ := rl.completer.GetBuffer()
+
+		rl.History.InsertMatch(line, cursor, true, forward, substring)
+
+		return
+	}
+
+	// Use the correct buffer for the rest of the function.
 	rl.line, rl.cursor, rl.selection = rl.completer.GetBuffer()
 
 	// Without multiline support, we always return the line.
 	if rl.AcceptMultiline == nil {
-		rl.display.AcceptLine()
-		rl.histories.Accept(hold, infer, nil)
+		keys := rl.Keys.Caller()
+		rl.Macros.StopRecord(keys)
+
+		rl.Display.AcceptLine()
+		rl.History.Accept(hold, infer, nil)
+
 		return
 	}
 
 	// Ask the caller if the line should be accepted
 	// as is, save the command line and accept it.
 	if rl.AcceptMultiline(*rl.line) {
-		rl.display.AcceptLine()
-		rl.histories.Accept(hold, infer, nil)
+		keys := rl.Keys.Caller()
+		rl.Macros.StopRecord(keys)
+
+		rl.Display.AcceptLine()
+		rl.History.Accept(hold, infer, nil)
+
 		return
 	}
 
@@ -454,14 +692,13 @@ func (rl *Shell) insertAutosuggestPartial(emacs bool) {
 		return
 	}
 
-	if !rl.opts.GetBool("history-autosuggest") {
+	if !rl.Config.GetBool("history-autosuggest") {
 		return
 	}
 
-	suggested := rl.histories.Suggest(rl.line)
-	if suggested.Len() > rl.line.Len() {
-		rl.undo.Save()
+	suggested := rl.History.Suggest(rl.line)
 
+	if suggested.Len() > rl.line.Len() {
 		var forward int
 
 		if emacs {
@@ -471,7 +708,7 @@ func (rl *Shell) insertAutosuggestPartial(emacs bool) {
 		}
 
 		if cpos+1+forward > suggested.Len() {
-			forward = suggested.Len() - cpos
+			forward = suggested.Len() - cpos - 1
 		}
 
 		rl.line.Insert(cpos+1, suggested[cpos+1:cpos+forward+1]...)
