@@ -7,6 +7,8 @@ import (
 	"fmt"
 	"io"
 	"os"
+	"os/user"
+	"path/filepath"
 	"strconv"
 	"strings"
 	"unicode"
@@ -16,6 +18,7 @@ import (
 // Parser is a inputrc parser.
 type Parser struct {
 	haltOnErr bool
+	strict    bool
 	name      string
 	app       string
 	term      string
@@ -191,6 +194,21 @@ func (p *Parser) doSet(h Handler, name, value string) error {
 	}
 	switch name {
 	case "keymap":
+		if p.strict {
+			switch value {
+			// see: man readline
+			// see: https://unix.stackexchange.com/questions/303479/what-are-readlines-modes-keymaps-and-their-default-bindings
+			case "emacs", "emacs-standard", "emacs-meta", "emacs-ctlx",
+				"vi", "vi-move", "vi-command", "vi-insert":
+			default:
+				return &ParseError{
+					Name: p.name,
+					Line: p.line,
+					Text: value,
+					Err:  ErrInvalidKeymap,
+				}
+			}
+		}
 		p.keymap = value
 		return nil
 	case "editing-mode":
@@ -279,7 +297,8 @@ func (p *Parser) do(h Handler, a, b string) error {
 		if !p.conds[len(p.conds)-1] {
 			return nil
 		}
-		buf, err := h.ReadFile(b)
+		path := expandIncludePath(b)
+		buf, err := h.ReadFile(path)
 		switch {
 		case err != nil && errors.Is(err, os.ErrNotExist):
 			return nil
@@ -310,6 +329,13 @@ type Option func(*Parser)
 func WithHaltOnErr(haltOnErr bool) Option {
 	return func(p *Parser) {
 		p.haltOnErr = haltOnErr
+	}
+}
+
+// WithStrict is a parser option to set strict keymap parsing.
+func WithStrict(strict bool) Option {
+	return func(p *Parser) {
+		p.strict = strict
 	}
 }
 
@@ -455,6 +481,7 @@ func decodeKey(r []rune, i, end int) (string, int, error) {
 	switch s {
 	case "":
 		return "", i, nil
+
 	case "delete", "del", "rubout":
 		c = Delete
 	case "escape", "esc":
@@ -578,8 +605,13 @@ func unescapeRunes(r []rune, i, end int) string {
 				}
 				i += 3
 			case c1 == 'M' && c2 == '-': // \M- meta prefix
-				s = append(s, Enmeta(c3))
-				i += 3
+				if c3 == 0 {
+					s = append(s, Esc)
+					i += 2
+				} else {
+					s = append(s, Enmeta(c3))
+					i += 3
+				}
 			default:
 				s = append(s, c1)
 				i += 1
@@ -610,4 +642,18 @@ func hexVal(c rune) rune {
 		return c - 'A' + 10
 	}
 	return c - '0'
+}
+
+// expandIncludePath handles tilde home directory expansion in $include path directives.
+func expandIncludePath(file string) string {
+	if !strings.HasPrefix(file, "~/") {
+		return file
+	}
+
+	u, err := user.Current()
+	if err != nil || u == nil || u.HomeDir == "" {
+		return file
+	}
+
+	return filepath.Join(u.HomeDir, file[2:])
 }
