@@ -6,29 +6,38 @@ import (
 	"strings"
 
 	"github.com/reeflective/readline/internal/color"
-	"github.com/reeflective/readline/internal/term"
 )
 
 // Iterations manages iterations for commands.
 type Iterations struct {
-	times  string
-	active bool // Are we currently setting the iterations.
-	set    bool // Has the last command been an iteration one.
-	reset  bool // Did a command reset the iterations (eg. vi-cmd-mode)
+	times   string // Stores iteration value
+	active  bool   // Are we currently setting the iterations.
+	pending bool   // Has the last command been an iteration one (vi-pending style)
 }
 
-// Add adds a string which might be a digit or a negative sign.
+// Add accepts a string to be converted as an integer representing
+// the number of times some action should be performed.
+// The times parameter can also be a negative sign, in which case
+// the iterations value will be negative until those are reset.
 func (i *Iterations) Add(times string) {
 	if times == "" {
 		return
 	}
 
-	i.active = true
-	i.set = true
+	// Never accept non-digit values.
+	if _, err := strconv.Atoi(times); err != nil && times != "-" {
+		return
+	}
 
-	if times == "-" || strings.HasPrefix(times, "-") {
-		i.times = times
-	} else {
+	i.active = true
+	i.pending = true
+
+	switch {
+	case times == "-":
+		i.times = times + i.times
+	case strings.HasPrefix(times, "-"):
+		i.times = "-" + i.times + strings.TrimPrefix(times, "-")
+	default:
 		i.times += times
 	}
 }
@@ -39,8 +48,12 @@ func (i *Iterations) Get() int {
 	times, err := strconv.Atoi(i.times)
 
 	// Any invalid value is still one time.
-	if err != nil && times == -1 {
+	if err != nil && strings.HasPrefix(i.times, "-") {
+		times = -1
+	} else if err != nil && times == 0 {
 		times = 1
+	} else if times == 0 && strings.HasPrefix(i.times, "-") {
+		times = -1
 	}
 
 	// At least one iteration
@@ -58,32 +71,33 @@ func (i *Iterations) IsSet() bool {
 	return i.active
 }
 
+// IsPending returns true if the very last command executed was an
+// iteration one. This is only meant for the main readline loop/run.
+func (i *Iterations) IsPending() bool {
+	return i.pending
+}
+
 // Reset resets the iterations (drops them).
 func (i *Iterations) Reset() {
 	i.times = ""
 	i.active = false
-	i.reset = true
-	i.set = false
+	i.pending = false
 }
 
 // Reset resets the iterations if the last command was not one to set them.
 // If the reset operated on active iterations, this function returns true.
-func (i *Iterations) ResetPostCommand() (hint string, wasActive bool) {
-	if i.active {
-		termWidth := term.GetWidth()
-		padLen := termWidth - len(i.times)
-		hint = color.Dim + fmt.Sprintf("%s%s", strings.Repeat(" ", padLen), i.times)
+func ResetPostRunIterations(iter *Iterations) (hint string) {
+	if iter.pending {
+		hint = color.Dim + fmt.Sprintf("(arg: %s)", iter.times)
 	}
 
-	if i.set {
-		i.set = false
+	if iter.pending {
+		iter.pending = false
+
 		return
 	}
 
-	wasActive = i.active || i.reset
-	i.times = ""
-	i.active = false
-	i.reset = false
+	iter.active = false
 
 	return
 }
