@@ -1,6 +1,9 @@
 package core
 
 import (
+	"regexp"
+	"unicode"
+
 	"github.com/reeflective/readline/inputrc"
 	"github.com/reeflective/readline/internal/color"
 	"github.com/reeflective/readline/internal/strutil"
@@ -11,16 +14,21 @@ import (
 // with the default cursor mark and position, and contains a list of additional surround
 // selections used to change/select multiple parts of the line at once.
 type Selection struct {
-	stype      string
-	active     bool
-	visual     bool
-	visualLine bool
-	bpos       int
-	epos       int
-	fg         string
-	bg         string
-	surrounds  []Selection
+	Type       string // Can be a normal one, surrounding (pairs), (cursor) matchers, etc.
+	active     bool   // The selection is running.
+	visual     bool   // The selection is highlighted.
+	visualLine bool   // The selection should span entire lines.
+	bpos       int    // Beginning index position
+	epos       int    // End index position (can be +1 in visual mode, to encompass cursor pos)
+	kpos       int    // Keyword regexp matchers cycling counter.
+	kmpos      int    // Keyword regexp matcher subgroups counter.
 
+	// Display
+	fg        string      // Foreground color of the highlighted selection.
+	bg        string      // Background color.
+	surrounds []Selection // Surrounds are usually pairs of characters matching each other (quotes/brackets, etc.)
+
+	// Core
 	line   *Line
 	cursor *Cursor
 }
@@ -56,7 +64,7 @@ func (s *Selection) MarkRange(bpos, epos int) {
 		return
 	}
 
-	s.stype = "visual"
+	s.Type = "visual"
 	s.active = true
 	s.bpos = bpos
 	s.epos = epos
@@ -67,13 +75,15 @@ func (s *Selection) MarkRange(bpos, epos int) {
 // The first area starts at bpos, and the second one at epos. If either bpos
 // is negative or epos is > line.Len()-1, no selection is created.
 func (s *Selection) MarkSurround(bpos, epos int) {
-	if bpos < 0 || epos > s.line.Len() {
+	if bpos < 0 || epos > s.line.Len()-1 {
 		return
 	}
 
+	s.active = true
+
 	for _, pos := range []int{bpos, epos} {
 		s.surrounds = append(s.surrounds, Selection{
-			stype:  "surround",
+			Type:   "surround",
 			active: true,
 			visual: true,
 			bpos:   pos,
@@ -93,7 +103,7 @@ func (s *Selection) Active() bool {
 }
 
 // Visual sets the selection as a visual one (highlighted),
-// which is core.y seen in Vim.
+// which is commonly seen in Vim edition mode.
 // If line is true, the visual is extended to entire lines.
 func (s *Selection) Visual(line bool) {
 	s.visual = true
@@ -119,7 +129,7 @@ func (s *Selection) Pos() (bpos, epos int) {
 	}
 
 	// Use currently set values, or update if one is pending.
-	bpos, epos = s.bpos, s.epos
+	s.bpos, s.epos = bpos, epos
 
 	if epos == -1 {
 		bpos, epos = s.selectToCursor(bpos)
@@ -129,8 +139,10 @@ func (s *Selection) Pos() (bpos, epos int) {
 		epos++
 	}
 
-	// Always check that neither the initial values
-	// nor the ones that we might have updated are wrong.
+	// Always check that neither the initial values nor the ones
+	// that we might have updated are wrong. It's very rare that
+	// the adjusted values would be invalid as a result of this
+	// call (unfixable values), but better being too safe than not.
 	bpos, epos, valid = s.checkRange(bpos, epos)
 	if !valid {
 		return -1, -1
@@ -266,17 +278,12 @@ func (s *Selection) InsertAt(bpos, epos int) {
 		return
 	}
 
+	// Get and reset the selection.
 	defer s.Reset()
 
-	// Get and reset the selection.
 	buf := s.Text()
-	if len(buf) == 0 {
-		return
-	}
 
 	switch {
-	case bpos == -1:
-		s.line.Insert(epos, []rune(buf)...)
 	case epos == -1, bpos == epos:
 		s.line.Insert(bpos, []rune(buf)...)
 	default:
@@ -313,8 +320,10 @@ func (s *Selection) Surround(bchar, echar rune) {
 // SelectAWord selects a word around the current cursor position,
 // selecting leading or trailing spaces depending on where the cursor
 // is: if on a blank space, in a word, or at the end of the line.
-func (s *Selection) SelectAWord() {
-	var bpos, epos int
+func (s *Selection) SelectAWord() (bpos, epos int) {
+	if s.line.Len() == 0 {
+		return
+	}
 
 	bpos = s.cursor.Pos()
 	cpos := bpos
@@ -325,7 +334,7 @@ func (s *Selection) SelectAWord() {
 	s.cursor.Set(epos)
 	cpos = s.cursor.Pos()
 
-	spaceAfter := cpos < s.line.Len()-1 && (*s.line)[cpos+1] == inputrc.Space
+	spaceAfter := cpos < s.line.Len()-1 && isSpace((*s.line)[cpos+1])
 
 	// And only select spaces after it if the word selected is not preceded
 	// by spaces as well, or if we started the selection within this word.
@@ -334,12 +343,18 @@ func (s *Selection) SelectAWord() {
 	if !s.Active() || bpos < cpos {
 		s.Mark(bpos)
 	}
+
+	return bpos, epos
 }
 
 // SelectABlankWord selects a bigword around the current cursor position,
 // selecting leading or trailing spaces depending on where the cursor is:
 // if on a blank space, in a word, or at the end of the line.
 func (s *Selection) SelectABlankWord() (bpos, epos int) {
+	if s.line.Len() == 0 {
+		return
+	}
+
 	bpos = s.cursor.Pos()
 	spaceBefore, spaceUnder := s.spacesAroundWord(bpos)
 
@@ -354,7 +369,7 @@ func (s *Selection) SelectABlankWord() (bpos, epos int) {
 
 	// Then go to the end of the blank word
 	s.cursor.Move(s.line.ForwardEnd(s.line.TokenizeSpace, s.cursor.Pos()))
-	spaceAfter := s.cursor.Pos() < s.line.Len()-1 && (*s.line)[s.cursor.Pos()+1] == inputrc.Space
+	spaceAfter := s.cursor.Pos() < s.line.Len()-1 && isSpace((*s.line)[s.cursor.Pos()+1])
 
 	// And only select spaces after it if the word selected is not preceded
 	// by spaces as well, or if we started the selection within this word.
@@ -371,6 +386,10 @@ func (s *Selection) SelectABlankWord() (bpos, epos int) {
 // selecting leading or trailing spaces depending on where the cursor
 // is: if on a blank space, in a word, or at the end of the line.
 func (s *Selection) SelectAShellWord() (bpos, epos int) {
+	if s.line.Len() == 0 {
+		return
+	}
+
 	s.cursor.CheckCommand()
 	s.cursor.ToFirstNonSpace(true)
 
@@ -389,11 +408,12 @@ func (s *Selection) SelectAShellWord() (bpos, epos int) {
 	// The quotes might be followed by non-blank characters,
 	// in which case we must keep expanding our selection.
 	for {
-		spaceBefore := mark > 0 && (*s.line)[mark-1] == inputrc.Space
+		spaceBefore := mark > 0 && isSpace((*s.line)[mark-1])
 		if spaceBefore {
 			s.cursor.Dec()
 			s.cursor.ToFirstNonSpace(false)
 			s.cursor.Inc()
+
 			break
 		} else if mark == 0 {
 			break
@@ -408,7 +428,7 @@ func (s *Selection) SelectAShellWord() (bpos, epos int) {
 
 	// Adjust if no spaces after.
 	for {
-		spaceAfter := cpos < s.line.Len()-1 && (*s.line)[cpos+1] == inputrc.Space
+		spaceAfter := cpos < s.line.Len()-1 && isSpace((*s.line)[cpos+1])
 		if spaceAfter || cpos == s.line.Len()-1 {
 			break
 		}
@@ -423,6 +443,43 @@ func (s *Selection) SelectAShellWord() (bpos, epos int) {
 	}
 
 	return bpos, cpos
+}
+
+// SelectKeyword attempts to find a pattern in the current blank word
+// around the current cursor position, using various regular expressions.
+// Repeatedly calling this function will cycle through all regex matches,
+// or if a matcher captured multiple subgroups, through each of those groups.
+//
+// Those are, in the order in which they are tried:
+// URI / URL / Domain|IPv4|IPv6 / URL path component / URL parameters.
+//
+// The returned positions are the beginning and end positions of the match
+// on the line (absolute position, not relative to cursor), or if no matcher
+// succeeds, the bpos and epos parameters are returned unchanged.
+// If found is true, it means a match occurred, otherwise false is returned.
+func (s *Selection) SelectKeyword(bpos, epos int, next bool) (kbpos, kepos int, match bool) {
+	if s.line.Len() == 0 {
+		return bpos, epos, false
+	}
+
+	selection := (*s.line)[bpos:epos]
+
+	_, match, kbpos, kepos = s.matchKeyword(selection, bpos, next)
+	if !match {
+		return bpos, epos, false
+	}
+
+	// Always check the validity of the selection
+	kbpos, kepos, match = s.checkRange(kbpos, kepos)
+	if !match {
+		return bpos, epos, false
+	}
+
+	// Mark the selection at its beginning
+	// and move the cursor to its end.
+	s.Mark(kbpos)
+
+	return kbpos, kepos, true
 }
 
 // ReplaceWith replaces all characters of the line within the current
@@ -457,14 +514,25 @@ func (s *Selection) Cut() (buf string) {
 
 	defer s.Reset()
 
-	bpos, epos := s.Pos()
-	if bpos == -1 || epos == -1 {
-		return
+	switch {
+	case len(s.surrounds) > 0:
+		offset := 0
+
+		for _, surround := range s.surrounds {
+			s.line.CutRune(surround.bpos - offset)
+			offset++
+		}
+
+	default:
+		bpos, epos := s.Pos()
+		if bpos == -1 || epos == -1 {
+			return
+		}
+
+		buf = s.Text()
+
+		s.line.Cut(bpos, epos)
 	}
-
-	buf = s.Text()
-
-	s.line.Cut(bpos, epos)
 
 	return
 }
@@ -479,17 +547,84 @@ func (s *Selection) Highlights() (fg, bg string) {
 	return s.fg, s.bg
 }
 
+// HighlightMatchers adds highlighting to matching
+// parens when the cursor is on one of them.
+func HighlightMatchers(sel *Selection) {
+	cpos := sel.cursor.Pos()
+
+	if sel.line.Len() == 0 || cpos == sel.line.Len() {
+		return
+	}
+
+	if strutil.IsBracket(sel.cursor.Char()) {
+		var adjust, ppos int
+
+		split, index, pos := sel.line.TokenizeBlock(cpos)
+
+		switch {
+		case len(split) == 0:
+			return
+		case pos == 0 && len(split) > index:
+			adjust = len(split[index])
+		default:
+			adjust = pos * -1
+		}
+
+		ppos = cpos + adjust
+
+		sel.surrounds = append(sel.surrounds, Selection{
+			Type:   "matcher",
+			active: true,
+			visual: true,
+			bpos:   ppos,
+			epos:   ppos,
+			bg:     color.SGR("240", false),
+			line:   sel.line,
+			cursor: sel.cursor,
+		})
+	}
+}
+
+// ResetMatchers is used by the display engine
+// to reset matching parens highlighting regions.
+func ResetMatchers(sel *Selection) {
+	var surrounds []Selection
+
+	for _, surround := range sel.surrounds {
+		if surround.Type == "matcher" {
+			continue
+		}
+
+		surrounds = append(surrounds, surround)
+	}
+
+	sel.surrounds = surrounds
+}
+
 // Reset makes the current selection inactive, resetting all of its values.
 func (s *Selection) Reset() {
-	s.stype = ""
+	s.Type = ""
 	s.active = false
 	s.visual = false
 	s.visualLine = false
 	s.bpos = -1
 	s.epos = -1
+	s.kpos = 0
 	s.fg = ""
 	s.bg = ""
-	s.surrounds = make([]Selection, 0)
+
+	// Get rid of all surround regions but matcher ones.
+	surrounds := make([]Selection, 0)
+
+	for _, surround := range s.surrounds {
+		if surround.Type != "matcher" {
+			continue
+		}
+
+		surrounds = append(surrounds, surround)
+	}
+
+	s.surrounds = surrounds
 }
 
 func (s *Selection) checkRange(bpos, epos int) (int, int, bool) {
@@ -572,15 +707,19 @@ func (s *Selection) selectToCursor(bpos int) (int, int) {
 }
 
 func (s *Selection) spacesAroundWord(cpos int) (before, under bool) {
-	under = (*s.line)[cpos] == inputrc.Space
-	before = cpos > 0 && (*s.line)[cpos-1] == inputrc.Space
+	under = isSpace(s.cursor.Char())
+	before = cpos > 0 && isSpace((*s.line)[cpos-1])
 
 	return
 }
 
+func isSpace(char rune) bool {
+	return unicode.IsSpace(char) && char != inputrc.Newline
+}
+
 // adjustWordSelection adjust the beginning and end of a word (blank or not) selection, depending
 // on whether it's surrounded by spaces, and if selection started from a whitespace or within word.
-func (s *Selection) adjustWordSelection(before, under, after bool, bpos int) (int, int) {
+func (s *Selection) adjustWordSelection(_, under, after bool, bpos int) (int, int) {
 	var epos int
 
 	if after && !under {
@@ -600,3 +739,131 @@ func (s *Selection) adjustWordSelection(before, under, after bool, bpos int) (in
 
 	return bpos, epos
 }
+
+func (s *Selection) matchKeyword(buf []rune, bbpos int, next bool) (name string, found bool, bpos, epos int) {
+	matchersNames := []string{
+		"URI",
+	}
+
+	matchers := map[string]*regexp.Regexp{
+		"URI": regexp.MustCompile(`https?:\/\/(?:www\.)?([-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b)*(\/[\/\d\w\.-]*)*(?:[\?])*(.+)*`),
+	}
+
+	// The matcher name and all the capturing subgroups it found.
+	var matcherName string
+	var groups []int
+
+	// Always preload the current/last target matcher used.
+	if s.kpos > 0 && s.kpos <= len(matchersNames) {
+		matcherName = matchersNames[s.kpos-1]
+		matcher := matchers[matcherName]
+
+		groups = s.runMatcher(buf, matcher)
+	}
+
+	// Either increment/decrement the counter (switch the matcher altogether),
+	// or cycle through the match groups for the current matcher.
+	if sbpos, sepos, cycled := s.cycleSubgroup(groups, bbpos, next); cycled {
+		return matcherName, true, sbpos, sepos
+	}
+
+	// We did not cycle through subgroups, so cycle the matchers.
+	prevPos := s.kpos
+	s.kmpos = 1
+
+	if next {
+		s.kpos++
+	} else {
+		s.kpos--
+	}
+
+	if s.kpos > len(matchersNames) {
+		s.kpos = 1
+	}
+
+	if s.kpos < 1 {
+		s.kpos = len(matchersNames)
+	}
+
+	// Prepare the cycling functions, increments and counters.
+	var (
+		kpos = s.kpos
+		done func(i int) bool
+		move func(inc int) int
+	)
+
+	if next {
+		done = func(i int) bool { return i <= len(matchersNames) }
+		move = func(inc int) int { return kpos + 1 }
+	} else {
+		done = func(i int) bool { return i > 0 }
+		move = func(inc int) int { return kpos - 1 }
+	}
+
+	// Try the different matchers until one succeeds, and select the first/last capturing group.
+	for done(kpos) {
+		matcherName = matchersNames[kpos-1]
+		matcher := matchers[matcherName]
+
+		groups = s.runMatcher(buf, matcher)
+
+		if len(groups) <= 1 {
+			kpos = move(kpos)
+			continue
+		}
+
+		if !next && prevPos == 1 {
+			s.kmpos = len(groups)
+
+			return matcherName, true, bbpos + groups[len(groups)-2], bbpos + groups[len(groups)-1]
+		}
+
+		return matcherName, true, bbpos + groups[0], bbpos + groups[1]
+	}
+
+	return "", false, -1, -1
+}
+
+func (s *Selection) cycleSubgroup(groups []int, bbpos int, next bool) (bpos, epos int, cycled bool) {
+	canCycleSubgroup := (next && s.kmpos < len(groups)-1) || (!next && s.kmpos > 1)
+
+	if !canCycleSubgroup {
+		return
+	}
+
+	if next {
+		s.kmpos++
+	} else {
+		s.kmpos--
+	}
+
+	return bbpos + groups[s.kmpos-1], bbpos + groups[s.kmpos], true
+}
+
+func (s *Selection) runMatcher(buf []rune, matcher *regexp.Regexp) (groups []int) {
+	if matcher == nil {
+		return
+	}
+
+	matches := matcher.FindAllStringSubmatchIndex(string(buf), -1)
+	if len(matches) == 0 {
+		return
+	}
+
+	return matches[0]
+}
+
+// Tested URL / IP regexp matchers, not working as well as the current ones
+//
+// "URL": regexp.MustCompile(`([\w+]+\:\/\/)?([\w\d-]+\.)*[\w-]+[\.\:]\w+([\/\?\=\&\#.]?[\w-]+)*\/?`),
+
+// "Domain|IP": regexp.MustCompile(ipv4_regex + `|` + ipv6_regex + `|` + domain_regex),
+// (http(s?):\/\/)?(((www\.)?+[a-zA-Z0-9\.\-\_]+(\.[a-zA-Z]{2,3})+)|(\b(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\b))(\/[a-zA-Z0-9\_\-\s\.\/\?\%\#\&\=]*)?
+// "Domain|IP": regexp.MustCompile(`((www\.)?[a-zA-Z0-9\.\-\_]+(\.[a-zA-Z]{2,3})+)|(\b(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\b)`),
+
+// "URL path": ,
+// "URL parameters": regexp.MustCompile(`[(\?|\&)]([^=]+)\=([^&#]+)`),
+
+// ipv6Regex   = `^(([0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,7}:|([0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4}){1,2}|([0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}|([0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|([0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|:((:[0-9a-fA-F]{1,4}){1,7}|:)|fe80:(:[0-9a-fA-F]{0,4}){0,4}%[0-9a-zA-Z]{1,}|::(ffff(:0{1,4}){0,1}:){0,1}((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])|([0-9a-fA-F]{1,4}:){1,4}:((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9]))`
+// ipv4Regex   = `^(((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)(\.|$)){4})`
+// domainRegex = `^(?:[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?\.)+[a-z0-9][a-z0-9-]{0,61}[a-z0-9]`
