@@ -1,21 +1,23 @@
-package ui
+package display
 
 import (
 	"fmt"
 	"regexp"
 	"sort"
+	"strconv"
+	"strings"
 
 	"github.com/reeflective/readline/internal/color"
 	"github.com/reeflective/readline/internal/core"
 )
 
-// Highlight applies visual/selection highlighting to a line.
+// highlightLine applies visual/selection highlighting to a line.
 // The provided line might already have been highlighted by a user-provided
 // highlighter: this function accounts for any embedded color sequences.
-func Highlight(line []rune, selection core.Selection) string {
+func (e *Engine) highlightLine(line []rune, selection core.Selection) string {
 	// Sort regions and extract colors/positions.
 	sorted := sortHighlights(selection)
-	colors := getHighlights(line, sorted)
+	colors := e.getHighlights(line, sorted)
 
 	var highlighted string
 
@@ -29,9 +31,12 @@ func Highlight(line []rune, selection core.Selection) string {
 	}
 
 	// Finally, highlight comments using a regex.
-	// TODO: Replace # with configured comment sign
-	commentsMatch := regexp.MustCompile(`(^|\s)#.*`)
-	highlighted = commentsMatch.ReplaceAllString(highlighted, fmt.Sprintf("%s${0}%s", color.FgBlackBright, color.Reset))
+	comment := strings.Trim(e.opts.GetString("comment-begin"), "\"")
+	commentPattern := fmt.Sprintf(`(^|\s)%s.*`, comment)
+
+	if commentsMatch, err := regexp.Compile(commentPattern); err == nil {
+		highlighted = commentsMatch.ReplaceAllString(highlighted, fmt.Sprintf("%s${0}%s", color.FgBlackBright, color.Reset))
+	}
 
 	highlighted += color.Reset
 
@@ -49,30 +54,47 @@ func sortHighlights(vhl core.Selection) []core.Selection {
 		bpos = append(bpos, rbpos)
 	}
 
-	all = append(all, vhl)
-
 	if vhl.Active() && vhl.IsVisual() {
+		all = append(all, vhl)
 		vbpos, _ := vhl.Pos()
 		bpos = append(bpos, vbpos)
 	}
 
 	sort.Ints(bpos)
 
+	prevIsMatcher := false
+	prevPos := 0
+
 	for _, pos := range bpos {
 		for _, reg := range all {
 			bpos, _ := reg.Pos()
+			isMatcher := reg.Type == "matcher"
 
-			if bpos == pos && reg.Active() && reg.IsVisual() {
-				sorted = append(sorted, reg)
-				break
+			if bpos != pos || !reg.Active() || !reg.IsVisual() {
+				continue
 			}
+
+			// If we have both a matcher and a visual selection
+			// starting at the same position, then we might have
+			// just added the matcher, and we must "overwrite" it
+			// with the visual selection, so skip until we find it.
+			if prevIsMatcher && isMatcher && prevPos == pos {
+				continue
+			}
+
+			// Else the region is good to be used in that order.
+			sorted = append(sorted, reg)
+			prevIsMatcher = reg.Type == "matcher"
+			prevPos = bpos
+
+			break
 		}
 	}
 
 	return sorted
 }
 
-func getHighlights(line []rune, sorted []core.Selection) map[int][]rune {
+func (e *Engine) getHighlights(line []rune, sorted []core.Selection) map[int][]rune {
 	highlights := make(map[int][]rune)
 
 	// Find any highlighting already applied on the line,
@@ -120,8 +142,8 @@ func getHighlights(line []rune, sorted []core.Selection) map[int][]rune {
 		}
 
 		// Add new colors if any, and reset if some are done.
-		regions, posHl = hlAdd(regions, posHl, pos)
-		posHl = hlReset(regions, newHl, posHl)
+		regions, posHl = e.hlReset(regions, posHl, pos)
+		posHl = e.hlAdd(regions, newHl, posHl)
 
 		// Add to the line, with the raw index since
 		// we must take into account embedded colors.
@@ -133,38 +155,68 @@ func getHighlights(line []rune, sorted []core.Selection) map[int][]rune {
 	return highlights
 }
 
-func hlAdd(regions []core.Selection, line []rune, pos int) ([]core.Selection, []rune) {
+func (e *Engine) hlAdd(regions []core.Selection, newHl core.Selection, line []rune) []rune {
+	var (
+		fg, bg  string
+		matcher bool
+		hl      core.Selection
+	)
+
+	if newHl.Active() {
+		hl = newHl
+	} else if len(regions) > 0 {
+		hl = regions[len(regions)-1]
+	}
+
+	fg, bg = hl.Highlights()
+	matcher = hl.Type == "matcher"
+
+	// Update the highlighting with inputrc settings if any.
+	if bg != "" && !matcher {
+		background := strings.ReplaceAll(e.opts.GetString("active-region-start-color"), `\e`, "\x1b")
+		if bg, _ = strconv.Unquote(background); bg == "" {
+			bg = color.Reverse
+		}
+	}
+
+	// Add highlightings
+	line = append(line, []rune(bg)...)
+	line = append(line, []rune(fg)...)
+
+	return line
+}
+
+func (e *Engine) hlReset(regions []core.Selection, line []rune, pos int) ([]core.Selection, []rune) {
 	for i, reg := range regions {
 		_, epos := reg.Pos()
 		foreground, background := reg.Highlights()
+		matcher := reg.Type == "matcher"
 
-		if epos == pos {
+		if epos != pos {
+			continue
+		}
+
+		if i < len(regions)-1 {
 			regions = append(regions[:i], regions[i+1:]...)
+		} else {
+			regions = regions[:i]
+		}
 
-			if foreground != "" {
-				line = append(line, []rune(color.FgDefault)...)
-			}
+		if foreground != "" {
+			line = append(line, []rune(color.FgDefault)...)
+		}
 
-			if background != "" {
+		if background != "" {
+			background, _ := strconv.Unquote(e.opts.GetString("active-region-end-color"))
+			foreground := e.opts.GetString("active-region-start-color")
+
+			if background == "" && foreground == "" && !matcher {
+				line = append(line, []rune(color.ReverseReset)...)
+			} else {
 				line = append(line, []rune(color.BgDefault)...)
 			}
 		}
 	}
 
 	return regions, line
-}
-
-func hlReset(regions []core.Selection, newHl core.Selection, line []rune) []rune {
-	if newHl.Active() {
-		fg, bg := newHl.Highlights()
-		line = append(line, []rune(bg)...)
-		line = append(line, []rune(fg)...)
-	} else if len(regions) > 0 {
-		backHl := regions[len(regions)-1]
-		fg, bg := backHl.Highlights()
-		line = append(line, []rune(bg)...)
-		line = append(line, []rune(fg)...)
-	}
-
-	return line
 }
