@@ -9,8 +9,6 @@ import (
 	"os/exec"
 )
 
-const defaultEditor = "vi"
-
 // ErrStart indicates that the command to start the editor failed.
 var ErrStart = errors.New("failed to start editor")
 
@@ -19,16 +17,13 @@ var ErrStart = errors.New("failed to start editor")
 // temp directory under this name.
 // If the filetype is not empty and if the system editor supports it, the
 // file will be opened with the specified filetype passed to the editor.
-func EditBuffer(buf []rune, filename, filetype string) ([]rune, error) {
+func EditBuffer(buf []rune, filename, filetype string, emacs bool) ([]rune, error) {
 	name, err := writeToFile([]byte(string(buf)), filename)
 	if err != nil {
 		return buf, err
 	}
 
-	editor := os.Getenv("EDITOR")
-	if editor == "" {
-		editor = defaultEditor
-	}
+	editor := getSystemEditor(emacs)
 
 	args := []string{}
 	if filetype != "" {
