@@ -4,94 +4,30 @@ import "github.com/reeflective/readline/inputrc"
 
 // emacsKeys are the default keymaps in Emacs mode.
 var emacsKeys = map[string]inputrc.Bind{
-	"^M":          {Action: "accept-line"},
-	"^A":          {Action: "beginning-of-line"},
-	"^B":          {Action: "backward-char"},
-	"^D":          {Action: "delete-char-or-list"},
-	"^E":          {Action: "end-of-line"},
-	"^F":          {Action: "forward-char"},
-	"^G":          {Action: "send-break"},
-	"^I":          {Action: "expand-or-complete"},
-	"^K":          {Action: "kill-line"},
-	"^L":          {Action: "clear-screen"},
-	"^N":          {Action: "down-line-or-history"},
-	"^P":          {Action: "up-line-or-history"},
-	"^O":          {Action: "accept-line-and-down-history"},
-	"^R":          {Action: "history-incremental-search-backward"},
-	"^S":          {Action: "history-incremental-search-forward"},
-	"^T":          {Action: "transpose-chars"},
-	"^U":          {Action: "kill-whole-line"},
-	"^W":          {Action: "backward-kill-word"},
-	"^Y":          {Action: "yank"},
-	"^X^B":        {Action: "vi-match-bracket"},
-	"^X^E":        {Action: "vi-edit-command-line"},
-	"^X^K":        {Action: "kill-buffer"},
-	"^X^N":        {Action: "infer-next-history"},
-	"^X^O":        {Action: "overwrite-mode"},
-	"^X^U":        {Action: "undo"},
-	"^X^V":        {Action: "vi-cmd-mode"},
-	"^X^X":        {Action: "exchange-point-and-mark"},
-	"^Xr":         {Action: "history-incremental-search-backward"},
-	"^Xs":         {Action: "history-incremental-search-forward"},
-	"^Xu":         {Action: "undo"},
-	"^[^D":        {Action: "list-choices"},
-	"^[^G":        {Action: "send-break"},
-	"^[^H":        {Action: "backward-kill-word"},
-	"^[^I":        {Action: "self-insert-unmeta"},
-	"^[^J":        {Action: "self-insert-unmeta"},
-	"^[^L":        {Action: "clear-screen"},
-	"^[^M":        {Action: "self-insert-unmeta"},
-	"^[^[OA":      {Action: "history-search"},
-	"^[^[OB":      {Action: "menu-select"},
-	"^[^[[A":      {Action: "history-search"},
-	"^[^[[B":      {Action: "menu-select"},
-	"^[^_":        {Action: "copy-prev-word"},
-	"^@":          {Action: "set-mark-command"},
-	`^["`:         {Action: "quote-region"},
-	"^['":         {Action: "quote-line"},
-	"^[-":         {Action: "neg-argument"},
-	"^[<":         {Action: "beginning-of-buffer-or-history"},
-	"^[>":         {Action: "end-of-buffer-or-history"},
-	"^[B":         {Action: "backward-word"},
-	"^[C":         {Action: "capitalize-word"},
-	"^[D":         {Action: "kill-word"},
-	"^[F":         {Action: "forward-word"},
-	"^[L":         {Action: "down-case-word"},
-	"^[N":         {Action: "history-search-forward"},
-	"^[P":         {Action: "history-search-backward"},
-	"^[Q":         {Action: "push-line"},
-	"^[S":         {Action: "spell-word"},
-	"^[T":         {Action: "transpose-words"},
-	"^[U":         {Action: "up-case-word"},
-	"^[W":         {Action: "copy-region-as-kill"},
-	"^[[1;3A":     {Action: "history-search"},
-	"^[[1;3B":     {Action: "menu-select"},
-	"^[[1;5C":     {Action: "forward-word"},
-	"^[[1;5D":     {Action: "backward-word"},
-	"^[[200~":     {Action: "bracketed-paste"},
-	"^[[3;5~":     {Action: "kill-word"},
-	"^[[3~":       {Action: "delete-char"},
-	"^[[5~":       {Action: "history-search"},
-	"^[[6~":       {Action: "menu-select"},
-	"^[[A":        {Action: "up-line-or-search"},
-	"^[[B":        {Action: "down-line-or-select"},
-	"^[[C":        {Action: "forward-char"},
-	"^[[D":        {Action: "backward-char"},
-	"^[[Z":        {Action: "menu-select"},
-	"^[a":         {Action: "accept-and-hold"},
-	"^[b":         {Action: "backward-word"},
-	"^[c":         {Action: "capitalize-word"},
-	"^[d":         {Action: "kill-word"},
-	"^[f":         {Action: "forward-word"},
-	"^[m":         {Action: "copy-prev-shell-word"},
-	"^[n":         {Action: "history-search-forward"},
-	"^[p":         {Action: "history-search-backward"},
-	"^[t":         {Action: "transpose-words"},
-	"^[u":         {Action: "up-case-word"},
-	"^[w":         {Action: "kill-region"},
-	"^[|":         {Action: "vi-goto-column"},
-	"^?":          {Action: "backward-delete-char"},
-	"^[^?":        {Action: "backward-kill-word"},
-	"^_":          {Action: "undo"},
-	"(\x1b[0-9])": {Action: "digit-argument"},
+	unescape(`\C-D`):     {Action: "end-of-file"},
+	unescape(`\C-N`):     {Action: "down-line-or-history"},
+	unescape(`\C-P`):     {Action: "up-line-or-history"},
+	unescape(`\C-x\C-b`): {Action: "vi-match"},
+	unescape(`\C-x\C-e`): {Action: "edit-command-line"},
+	unescape(`\C-x\C-n`): {Action: "infer-next-history"},
+	unescape(`\C-x\C-o`): {Action: "overwrite-mode"},
+	unescape(`\C-x\C-v`): {Action: "vi-movement-mode"},
+	unescape(`\C-Xr`):    {Action: "reverse-search-history"},
+	unescape(`\C-Xs`):    {Action: "forward-search-history"},
+	unescape(`\C-Xu`):    {Action: "undo"},
+	unescape(`\M-\C-^`):  {Action: "copy-prev-word"},
+	unescape(`\M-'`):     {Action: "quote-line"},
+	unescape(`\M-<`):     {Action: "beginning-of-buffer-or-history"},
+	unescape(`\M->`):     {Action: "end-of-buffer-or-history"},
+	unescape(`\M-a`):     {Action: "accept-and-hold"},
+	unescape(`\M-c`):     {Action: "capitalize-word"},
+	unescape(`\M-d`):     {Action: "kill-word"},
+	unescape(`\M-m`):     {Action: "copy-prev-shell-word"},
+	unescape(`\M-n`):     {Action: "history-search-forward"},
+	unescape(`\M-p`):     {Action: "history-search-backward"},
+	unescape(`\M-u`):     {Action: "up-case-word"},
+	unescape(`\M-w`):     {Action: "kill-region"},
+	unescape(`\M-|`):     {Action: "vi-goto-column"},
+	unescape(`\M-[A`):    {Action: "up-line-or-search"},
+	unescape(`\M-[B`):    {Action: "down-line-or-select"},
 }
