@@ -2,10 +2,10 @@ package strutil
 
 import (
 	"strings"
-	"unicode/utf8"
 
 	"github.com/reeflective/readline/internal/color"
 	"github.com/reeflective/readline/internal/term"
+	"github.com/rivo/uniseg"
 )
 
 // FormatTabs replaces all '\t' occurrences in a string with 6 spaces each.
@@ -13,16 +13,18 @@ func FormatTabs(s string) string {
 	return strings.ReplaceAll(s, "\t", "     ")
 }
 
-// RealLength returns the real length of a string, trimming all ANSI escaped codes.
+// RealLength returns the real length of a string (the number of terminal
+// columns used to render the line, which may contain special graphemes).
+// Before computing the width, it replaces tabs with (4) spaces, and strips colors.
 func RealLength(s string) int {
-	colors := color.Strip(s)                          // Remove colors
-	tabs := strings.ReplaceAll(colors, "\t", "     ") // Remove spaces
+	colors := color.Strip(s)
+	tabs := strings.ReplaceAll(colors, "\t", "     ")
 
-	return utf8.RuneCountInString(tabs)
+	return uniseg.StringWidth(tabs)
 }
 
 // LineSpan computes the number of columns and lines that are needed for a given line,
-// accounting for any ANSI escapes/color codes and wrapping with the current term width.
+// accounting for any ANSI escapes/color codes, and tabulations replaced with 4 spaces.
 func LineSpan(line []rune, idx, indent int) (x, y int) {
 	termWidth := term.GetWidth()
 	lineLen := RealLength(string(line))
