@@ -1,7 +1,7 @@
 package readline
 
 import (
-	"os/user"
+	"fmt"
 
 	"github.com/reeflective/readline/inputrc"
 	"github.com/reeflective/readline/internal/completion"
@@ -11,6 +11,7 @@ import (
 	"github.com/reeflective/readline/internal/history"
 	"github.com/reeflective/readline/internal/keymap"
 	"github.com/reeflective/readline/internal/macro"
+	"github.com/reeflective/readline/internal/term"
 	"github.com/reeflective/readline/internal/ui"
 )
 
@@ -19,37 +20,35 @@ import (
 // captures without having to repeatedly unload configuration.
 type Shell struct {
 	// Core editor
-	keys       *core.Keys        // Keys read user input keys and manage them.
-	line       *core.Line        // The input line buffer and its management methods.
-	cursor     *core.Cursor      // The cursor and its medhods.
-	undo       *core.LineHistory // Line undo/redo history.
-	selection  *core.Selection   // The selection managees various visual/pending selections.
-	iterations *core.Iterations  // Digit arguments for repeating commands.
-	buffers    *editor.Buffers   // buffers (Vim registers) and methods use/manage/query them.
+	line       *core.Line       // The input line buffer and its management methods.
+	cursor     *core.Cursor     // The cursor and its medhods.
+	selection  *core.Selection  // The selection managees various visual/pending selections.
+	Iterations *core.Iterations // Digit arguments for repeating commands.
+	Buffers    *editor.Buffers  // buffers (Vim registers) and methods use/manage/query them.
+	Keys       *core.Keys       // Keys is in charge of reading and managing buffered user input.
+	Keymap     *keymap.Engine   // Manages main/local keymaps, binds, stores command functions, etc.
+	History    *history.Sources // History manages all history types/sources (past commands and undo)
+	Macros     *macro.Engine    // Record, use and display macros.
 
 	// User interface
-	opts      *inputrc.Config    // Contains all keymaps, binds and per-application settings.
-	prompt    *ui.Prompt         // The prompt engine computes and renders prompt strings.
-	hint      *ui.Hint           // Usage/hints for completion/isearch below the input line.
+	Config    *inputrc.Config    // Contains all keymaps, binds and per-application settings.
+	Opts      []inputrc.Option   // Inputrc file parsing options (app/term/values, etc).
+	Prompt    *ui.Prompt         // The prompt engine computes and renders prompt strings.
+	Hint      *ui.Hint           // Usage/hints for completion/isearch below the input line.
 	completer *completion.Engine // Completions generation and display.
-	histories *history.Sources   // All history sources and management methods.
-	macros    *macro.Engine      // Record, use and display macros.
-	display   *display.Engine    // Manages display refresh/update/clearing.
-
-	// Others
-	keymaps *keymap.Modes // Manages main/local keymaps and runs key matching.
+	Display   *display.Engine    // Manages display refresh/update/clearing.
 
 	// User-provided functions
-	//
+
 	// AcceptMultiline enables the caller to decide if the shell should keep reading
 	// for user input on a new line (therefore, with the secondary prompt), or if it
 	// should return the current line at the end of the `rl.Readline()` call.
 	// This function should return true if the line is deemed complete (thus asking
 	// the shell to return from its Readline() loop), or false if the shell should
-	// keep reading input.
+	// keep reading input on a newline (thus, insert a newline and read).
 	AcceptMultiline func(line []rune) (accept bool)
 
-	// SyntaxHighlight is a helper function to provide syntax highlighting.
+	// SyntaxHighlighter is a helper function to provide syntax highlighting.
 	// Once enabled, set to nil to disable again.
 	SyntaxHighlighter func(line []rune) string
 
@@ -60,12 +59,12 @@ type Shell struct {
 
 	// SyntaxCompletion is used to autocomplete code syntax (like braces and
 	// quotation marks). If you want to complete words or phrases then you might
-	// be better off using the TabCompletion function.
+	// be better off using the Completer function.
 	// SyntaxCompletion takes the line ([]rune) and cursor position, and returns
 	// the new line and cursor position.
 	SyntaxCompleter func(line []rune, cursor int) ([]rune, int)
 
-	// HintText is a helper function which displays hint text the prompt.
+	// HintText is a helper function which displays hint text below the line.
 	// HintText takes the line input from the promt and the cursor position.
 	// It returns the hint text to display.
 	HintText func(line []rune, cursor int) []rune
@@ -73,80 +72,117 @@ type Shell struct {
 
 // NewShell returns a readline shell instance initialized with a default
 // inputrc configuration and binds, and with an in-memory command history.
-func NewShell() *Shell {
+// The constructor accepts an optional list of inputrc configuration options,
+// which are used when parsing/loading and applying any inputrc configuration.
+func NewShell(opts ...inputrc.Option) *Shell {
 	shell := new(Shell)
-
-	// Configuration
-	opts := inputrc.NewDefaultConfig()
-	if user, err := user.Current(); err == nil {
-		inputrc.UserDefault(user, opts)
-	}
 
 	// Core editor
 	keys := new(core.Keys)
 	line := new(core.Line)
 	cursor := core.NewCursor(line)
-	undo := core.NewLineHistory(line, cursor)
 	selection := core.NewSelection(line, cursor)
 	iterations := new(core.Iterations)
 
-	shell.keys = keys
+	shell.Keys = keys
 	shell.line = line
 	shell.cursor = cursor
 	shell.selection = selection
-	shell.undo = undo
-	shell.buffers = editor.NewBuffers()
-	shell.iterations = iterations
+	shell.Buffers = editor.NewBuffers()
+	shell.Iterations = iterations
 
 	// Keymaps and commands
-	keymaps := keymap.NewModes(keys, iterations, opts)
+	keymaps, config := keymap.NewEngine(keys, iterations, opts...)
 	keymaps.Register(shell.standardCommands())
 	keymaps.Register(shell.viCommands())
 	keymaps.Register(shell.historyCommands())
 	keymaps.Register(shell.completionCommands())
 
-	shell.keymaps = keymaps
+	shell.Keymap = keymaps
+	shell.Config = config
+	shell.Opts = opts
 
 	// User interface
 	hint := new(ui.Hint)
-	prompt := ui.NewPrompt(keys, line, cursor, keymaps, opts)
+	prompt := ui.NewPrompt(line, cursor, keymaps, config)
 	macros := macro.NewEngine(keys, hint)
-	completer := completion.NewEngine(keys, line, cursor, selection, hint, keymaps, opts)
-	completer.SetAutocompleter(shell.commandCompletion)
-	history := history.NewSources(line, cursor, hint)
-	display := display.NewEngine(keys, selection, history, prompt, hint, completer, opts)
+	history := history.NewSources(line, cursor, hint, config)
+	completer := completion.NewEngine(hint, keymaps, config)
+	completion.Init(completer, keys, line, cursor, selection, shell.commandCompletion)
 
-	shell.opts = opts
-	shell.hint = hint
-	shell.prompt = prompt
+	display := display.NewEngine(keys, selection, history, prompt, hint, completer, config)
+
+	shell.Config = config
+	shell.Hint = hint
+	shell.Prompt = prompt
 	shell.completer = completer
-	shell.macros = macros
-	shell.histories = history
-	shell.display = display
+	shell.Macros = macros
+	shell.History = history
+	shell.Display = display
 
 	return shell
 }
 
-// init gathers all steps to perform at the beginning of readline loop.
-func (rl *Shell) init() {
-	// Some components need the last accepted line.
-	rl.histories.Init()
+// Line is the shell input line buffer.
+// Contains methods to search and modify its contents,
+// split itself with tokenizers, and displaying itself.
+//
+// When the shell is in incremental-search mode, this line is the minibuffer.
+func (rl *Shell) Line() *core.Line { return rl.line }
 
-	// Reset core editor components.
-	rl.selection.Reset()
-	rl.buffers.Reset()
-	rl.undo.Reset()
-	rl.keys.Flush()
-	rl.cursor.ResetMark()
-	rl.cursor.Set(0)
-	rl.line.Set([]rune{}...) // TODO: Wrong; if line was inferred this resets it while it should not.
-	rl.undo.Save()
+// Cursor is the cursor position in the current line buffer.
+// Contains methods to set, move, describe and check itself.
+//
+// When the shell is in incremental-search mode, this cursor is the minibuffer's one.
+func (rl *Shell) Cursor() *core.Cursor { return rl.cursor }
 
-	// Reset/initialize user interface components.
-	rl.hint.Reset()
-	rl.completer.Cancel(true, true)
-	rl.display.Init(rl.SyntaxHighlighter)
+// Selection contains all regions of an input line that are currently selected/marked
+// with either a begin and/or end position. The main selection is the visual one, used
+// with the default cursor mark and position, and contains a list of additional surround
+// selections used to change/select multiple parts of the line at once.
+func (rl *Shell) Selection() *core.Selection { return rl.selection }
 
-	// Reset other components.
-	rl.iterations.Reset()
+// Printf prints a formatted string below the current line and redisplays the prompt
+// and input line (and possibly completions/hints if active) below the logged string.
+// A newline is added to the message so that the prompt is correctly refreshed below.
+func (rl *Shell) Printf(msg string, args ...any) (n int, err error) {
+	// First go back to the last line of the input line,
+	// and clear everything below (hints and completions).
+	rl.Display.CursorBelowLine()
+	term.MoveCursorBackwards(term.GetWidth())
+	fmt.Print(term.ClearScreenBelow)
+
+	// Skip a line, and print the formatted message.
+	n, err = fmt.Printf(msg+"\n", args...)
+
+	// Redisplay the prompt, input line and active helpers.
+	rl.Prompt.PrimaryPrint()
+	rl.Display.Refresh()
+
+	return
+}
+
+// PrintTransientf prints a formatted string in place of the current prompt and input
+// line, and then refreshes, or "pushes" the prompt/line below this printed message.
+func (rl *Shell) PrintTransientf(msg string, args ...any) (n int, err error) {
+	// First go back to the beginning of the line/prompt, and
+	// clear everything below (prompt/line/hints/completions).
+	// if rl.Prompt.Refreshing() {
+	// 	term.MoveCursorUp(1)
+	// }
+
+	rl.Display.CursorToLineStart()
+	term.MoveCursorBackwards(term.GetWidth())
+
+	term.MoveCursorUp(rl.Prompt.PrimaryUsed() + 1)
+	fmt.Print(term.ClearScreenBelow)
+
+	// Print the logged message.
+	n, err = fmt.Printf(msg+"\n", args...)
+
+	// Redisplay the prompt, input line and active helpers.
+	rl.Prompt.PrimaryPrint()
+	rl.Display.Refresh()
+
+	return
 }
